% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{abrvMonth}
\alias{abrvMonth}
\title{abrvMonth}
\usage{
abrvMonth(date)
}
\arguments{
\item{date}{A date value in Sys.Date or as.Date format.}
}
\value{
The month of the date in abbreviated format.
}
\description{
Returns the month in an abbreviated format of a date object.
}
\examples{

# Example abbreviating the current month today 
abrvMonth(Sys.Date())

# Example abbreviating the specified month
abrvMonth(as.Date('2024-11-23'))


}
\seealso{
\code{\link[=Month]{Month()}} and \code{\link[base:format]{base::format()}}
}
