% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessREDCap.R
\name{accessREDCap}
\alias{accessREDCap}
\title{Access REDCap data for country}
\usage{
accessREDCap(country, filepath = NULL)
}
\arguments{
\item{country}{A character string representing the country for which REDCap data will be accessed.}

\item{filepath}{Directory where extracted file should be saved}
}
\value{
If a secret key and file path exist, this function will proceed to call \code{downloadRedCap()} with the appropriate arguments.
Otherwise, it will prompt the user to create a secret key or file path, or to continue with the download without them.
}
\description{
This function checks for an existing secret key and file path for the specified country.
If these exist, the function will proceed to download REDCap data.
If not, the user will be prompted to create the necessary credentials.
}
\examples{
\dontrun{
# Request data for a fictional country 'DB'
accessREDCap("DB")
}
}
\seealso{
\code{\link{downloadRedCap}} for downloading REDCap data.

\code{\link{grabKeyring}} for creating or getting a keyring.

\code{\link{setOrGetFilePath}} for setting or getting a file path.
}
