% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{addImage}
\alias{addImage}
\title{addImage}
\usage{
addImage(
  onesheet,
  image,
  line,
  col,
  wide = 7,
  high = 4,
  unit = "in",
  spec.dpi = 300,
  wb = NULL
)
}
\arguments{
\item{onesheet}{A sheet object from openXlsx package}

\item{image}{The filename of a previously saved image (format supported: jpeg, bmp, png)
or the name of the image in the current R environment}

\item{line}{The line where to paste the image}

\item{col}{The column where to paste the image}

\item{wide}{Size of the image (in unit)}

\item{high}{Size of the image (in unit)}

\item{unit}{For wide and high, default is "in" (Inch) "cm" or "px"}

\item{spec.dpi}{300 by default}

\item{wb}{An optional wb if not already opened}
}
\value{
Nothing
}
\description{
Insert either an existing image or the current plot into a worksheet.
}
\examples{
\dontrun{ 
addImage(onesheet = cells, 
          image = graphname, 
          line = 2, 
          col = 2, 
          wb = template) 
}

# Usage 1: Insert existing image saved as a PNG file
xls_file <- externalFile("excelfile.xlsx")
openXlsx(xls_file)
cells <- openSheet("T2")
image_file <- externalFile("image_xlsx.png")
addImage(onesheet = cells, image = image_file, line = 20, col = 2)

# Usage 2: Insert current plot
\dontrun{
image_gg <- ggplot(Orange, 
                   aes(x = age, y = circumference, colour = as.factor(Tree))) +
                   geom_point() + 
                   geom_line() +
                   labs(colour = "Tree") + 
                   theme_minimal()
addImage(onesheet = cells, image = image_gg, line = 20, col = 2)
}
       
}
\seealso{
\code{\link[openxlsx:insertImage]{openxlsx::insertImage()}}, \code{\link[openxlsx:insertPlot]{openxlsx::insertPlot()}}
}
\author{
Jenny Howard, Marine Maurel
}
