% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blue}
\alias{blue}
\title{blue}
\usage{
blue(...)
}
\arguments{
\item{...}{list of values to be outputted in blue}
}
\value{
nothing
}
\description{
Send text colored in blue to the console
}
\details{
red, blue, italic and bold functions are used to decor text outputted to the console
syntax is similar to \code{cat()} syntax
}
\examples{
blue("text in blue")
}
