% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bold}
\alias{bold}
\title{bold}
\usage{
bold(...)
}
\arguments{
\item{...}{The text or values to be outputted in bold. Must be specified in "".}
}
\value{
nothing
}
\description{
This function is to make text and values appear in bold.
}
\examples{

# Example use of the function
bold("text in bold")



}
