% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bye}
\alias{bye}
\title{Say goodbye and stop execution}
\usage{
bye(message)
}
\arguments{
\item{message}{A character string to be printed as a goodbye message.}
}
\value{
No return value, called for side effects.
}
\description{
This function prints a goodbye message provided by the user and then
stop the execution of the R script.
}
\examples{
bye("Goodbye, see you later!")
}
