% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{calYear}
\alias{calYear}
\title{calYear}
\usage{
calYear(date)
}
\arguments{
\item{date}{A date value in the format Sys.Date or as.Date. Specified in "".}
}
\value{
The calendar year of the date as numeric.
}
\description{
Extract the year of a date object returned as numeric.
}
\examples{

# The year of the current date 
calYear(Sys.Date())

# The year of a specified date 
calYear(as.Date('2023-12-31'))

}
\seealso{
\code{\link[=isoYear]{isoYear()}} and \code{\link[base:format]{base::format()}}
}
