% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{camel}
\alias{camel}
\title{Set of utility functions used widely in the epiuf package}
\usage{
camel(x)
}
\arguments{
\item{x}{: A string containing "snake_case" or a string list or a vector of strings}
}
\value{
x with any occurrence transformed in to a CamelCase
}
\description{
Transform snake_case into CamelCase
}
\details{
Camel can be used to transform snake_case notation into a CamelCase notation.
You can use it on string, string list or string vector. This could be usefull to transform
all columns name into a standard notation
}
\examples{
test <- "snake_case"
camel(test)

}
