% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charCount}
\alias{charCount}
\title{charCount
count number of specific char into a text using reg expr}
\usage{
charCount(pattern, stosearch)
}
\arguments{
\item{pattern}{"The character or pattern to searh}

\item{stosearch}{The string to search in}
}
\value{
number of match
}
\description{
charCount
count number of specific char into a text using reg expr
}
\examples{
nb <- charCount("/", "test/essai/try")

}
