% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDico.R
\name{checkDico}
\alias{checkDico}
\title{checkDico}
\usage{
checkDico(data, varname, dicoCode)
}
\arguments{
\item{data}{Dataset which contain \code{varname}}

\item{varname}{The variable to chack against dico}

\item{dicoCode}{A dico code list}
}
\value{
text outputs
}
\description{
checkDico
}
