% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDicoAll.R
\name{checkDicoAll}
\alias{checkDicoAll}
\title{checkDicoAll}
\usage{
checkDicoAll(data, dictionary = NULL, dicos = NULL)
}
\arguments{
\item{data}{Dataset to process (data.frame)}

\item{dictionary}{An optional dictionary (epiuf strucure see \link{getDictionary})}

\item{dicos}{A Dico structure containing code and labels}
}
\value{
text outputs
}
\description{
Compares a data.frame to the contents of an optional dictionary. Flags a warning if the dictionary contains numbers not specified in the data.frame.
}
