% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmissing.R
\name{checkmissing}
\alias{checkmissing}
\title{checkMissing}
\usage{
checkmissing(data, varlist = NULL, sort = FALSE, showall = FALSE)
}
\arguments{
\item{data}{a data.frame, a vector or a vector of colnames}

\item{varlist}{a list of variable to be tested. Is used only if "what" is not a dataframe}

\item{sort}{boolean If TRUE the output table is sorted 'NOT IMPLEMENTED'}

\item{showall}{boolean if FALSE, the default, then variables with no missing values are
not included in the table}
}
\value{
list of 2 tables
}
\description{
checkMissing
}
\examples{
# still nothing
}
\author{
Gilles Desve
}
