% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanNames}
\alias{cleanNames}
\title{Clean names by removing special char and accented}
\usage{
cleanNames(name, chr = "")
}
\arguments{
\item{name}{A string to clean by removing non ascii char}

\item{chr}{A chr used to replace non ascii character}
}
\value{
the cleaned string
}
\description{
cleanNames can be used to obtain valid column names by removing non ascii char.
All invalid characters would be replaced by chr, an optional
character, if provided otherwise they are removed
In the same time, all accented character are replaced by a simple equivalent char
}
\examples{
x <- "ÁbcdêãçoàúüEssai/=+$67"
cleanNames(x)  # no libraries needed
}
