% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseVar.R
\name{collapseVar}
\alias{collapseVar}
\title{collapseVar}
\usage{
collapseVar(data, varname, hierarchy)
}
\arguments{
\item{data}{The dataset}

\item{varname}{Name of variable}

\item{hierarchy}{ordered list to use in collapse}
}
\value{
The variable collapsed and warnings if not possible
}
\description{
collapseVar
}
