% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseVarAll.R
\name{collapseVarAll}
\alias{collapseVarAll}
\title{collapseVarAll}
\usage{
collapseVarAll(data, action = NULL)
}
\arguments{
\item{data}{Data set to be processed (data.frame).}

\item{action}{A Dictionary action (list structure).}
}
\value{
The dataset with collapsed variable.
}
\description{
Applies a dictionary action to a data set.
}
