% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_drop.R
\name{comp}
\alias{comp}
\title{comp  >> Obsolete replaced by desc and desc by}
\usage{
comp(vars, comp.var, labels, data)
}
\arguments{
\item{vars}{list of vars you want to describe}

\item{comp.var}{variable to compare (need to be data$var)}

\item{labels}{labels for the list of vars}

\item{data}{name of the dataset}
}
\value{
the result
}
\description{
Aim : To describe numeric variables or categorical variables (need to be in factor)
according to an another variable
}
\examples{
data <- data.frame(Id = 1:4 , 
                    Vaccs = c("pfizer,moderna"," ", "pfizer", "moderna"),
                    Cov = 1,0,1,0)
vars <- c("Id", "Vaccs")
labels <- c("Id number", "Vaccination Brand")
\dontrun{
tab.comp1 <- comp(vars,data$Cov , labels, data) 
}
}
