% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDataSet.R
\name{compareDataSet}
\alias{compareDataSet}
\title{compareDataSet}
\usage{
compareDataSet(dfmod, dftest)
}
\arguments{
\item{dfmod}{A model dataset (one record at least)}

\item{dftest}{A dataset to test against the model}
}
\value{
a list of difference
}
\description{
Compare two dataset for difference in terms of structure
}
