% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confirm}
\alias{confirm}
\title{confirm
Confirmation dialog box}
\usage{
confirm(message = "")
}
\arguments{
\item{message}{The prompt to display before confirmation}
}
\value{
logical True if answer is "Yes"
}
\description{
confirm
Confirmation dialog box
}
