% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixencoding.R
\name{convEncoding}
\alias{convEncoding}
\title{convEncoding}
\usage{
convEncoding(df, originalEncoding = "UTF-8", targetEncoding = NULL)
}
\arguments{
\item{df}{A dataset}

\item{originalEncoding}{The originale encoding value (accepted value can be obtained via iconvlist())}

\item{targetEncoding}{The target encoding value (For R should be latin1 or UTF-8)}
}
\value{

}
\description{
convert character column in a dataset to a new encoding
(old encoding should be given)
}
