% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyTo.R
\name{copyTo}
\alias{copyTo}
\title{copyTo}
\usage{
copyTo(
  files_to_copy,
  destination,
  date_stamp = TRUE,
  commit = FALSE,
  commit_msg = NULL,
  pull = FALSE,
  push = FALSE
)
}
\arguments{
\item{files_to_copy}{List of filepaths for all files to be copied}

\item{date_stamp}{Logical - do you want to create a date of backup marker in reference backup folder}

\item{commit}{Logical - do you want to make a commit of your repository before extraction?}

\item{commit_msg}{If commit, what is your message. Default is: commit pre zip (todays date)}

\item{pull}{Logical - do you want to make a pull from github before extraction?}

\item{push}{Logical - do you want to make a push to github before extraction?}

\item{desitation}{Filepath for where you want to save the copied files}
}
\description{
Description
Wrapper function for file.copy which takes list of files specified and copies
to a desired folder.
For use when making backups of file: if date_stamp is
TRUE, a text document is created in the specified folder called
'date_last_copy.txt' with todays date written inside.
For use when saving copied files to a local github reposity: options to make a comit
and pull from then push to github after copying the files.
Utility in data management: retrieve refernce documents used by scripts which
are external to github and create a backup within the github repository.
}
\seealso{
For more details see the link below to access the vignette:
\href{../doc/epiuf_package.html}{\code{vignette("epiuf_package")}}
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
