% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{countIsoWeeks}
\alias{countIsoWeeks}
\title{countIsoWeeks}
\usage{
countIsoWeeks(date, origin = "2020-10-05")
}
\arguments{
\item{date}{A date value in the format of either Sys.Date or as.Date.}

\item{origin}{A reference date value in character string format.}
}
\value{
The number of weeks between the two dates.
}
\description{
Return the number of weeks between two dates.
}
\examples{

# Weeks difference between today's date and another
countIsoWeeks(date = Sys.Date(), origin = "2000-01-01")

# Weeks difference between two specified dates 
countIsoWeeks(date = "2021-01-01", origin = "2020-12-10")

}
\seealso{
\code{\link[base:difftime]{base::difftime()}}
}
\author{
Jenny Howard \email{j.howard@epiconcept.fr}
}
