% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring.r
\name{createKeyring}
\alias{createKeyring}
\title{Create a new secret in default Keyring}
\usage{
createKeyring(keyring_id, secret = NULL)
}
\arguments{
\item{keyring_id}{A string specifying the keyring ID.}

\item{secret}{A secret string to be stored. It is not recommended to use this parameters because in that case your secret
will be visible in your code. This parameters should be used only for tests}
}
\value{
A string with the keyring password.
}
\description{
This function creates a new secret with a specified keyring ID.
If the keyring already exists, it prompts to use ModifyKeyring instead.
Notice: the keyringid is postfixed with _epiufkeyring to avoid conflicts
}
