% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateOrGrabKeyring.r
\name{CreateOrGrabKeyring}
\alias{CreateOrGrabKeyring}
\title{Create or Grab a Keyring}
\usage{
CreateOrGrabKeyring(keyring_id)
}
\arguments{
\item{keyring_id}{A character string specifying the keyring name to check or create.}
}
\value{
A character string containing the password for further API calls.
}
\description{
This function checks if the specified keyring exists, and if not, it creates a new keyring with the provided password.
If the keyring exists, it uses the existing password for further API calls.
}
\examples{
\dontrun{
# Set a secret for a fictional country called 'DB'
CreateOrGrabKeyring("DB")
}
}
