% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasourcestools.R
\name{createSourceList}
\alias{createSourceList}
\title{Title}
\usage{
createSourceList(filename)
}
\arguments{
\item{filename}{Name of the Excel file containing source information}
}
\value{
nothing
}
\description{
Title
}
\examples{
\dontrun{ createSourceList("datasources.xls") }
}
