% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCheck.R
\name{dataCheck}
\alias{dataCheck}
\title{dataCheck}
\usage{
dataCheck(
  data,
  data_old = NULL,
  cond = "",
  text = "",
  varname = NULL,
  threshold = NULL
)
}
\arguments{
\item{data}{(data.frame) The dataset to check.}

\item{data_old}{(data.frame) A previously checked dataset. This dataset is optional
and can include the new dataset to be checked \code{'data'}. This option is useful
when we aim to check a dataset that has been appended with new data.
This will allow to identify in the new dataset, which observations have already
been identified and checked.}

\item{cond}{(character) The logical condition(s) to test,
specified as character string (e.g., \code{"Age > 120"}).}

\item{text}{(character) The message specified to prefix the returned unique
identifiers meeting the given condition (e.g., \code{"Nb and list of IDs with invalid Age value"}).}

\item{varname}{(character) The name of the variable in dataset \code{'data'}
including unique identifiers. It will be used to identify and return the list of observations meeting
the given condition \code{'cond'} (e.g., \code{"patientID"}).}

\item{threshold}{(numeric) The maximum number of unique identifiers to display in the output (default: 50).}
}
\value{
dataCheck returns a vector of 3 character string elements:
\itemize{
\item The message (\code{text}) associated with the tested condition (e.g., \code{"Nb and list of IDs with invalid Age value"})
\item The number of rows meeting the condition (e.g., \code{"3"}), and
\item The list of unique identifiers from \code{varname} of the rows
meeting the given condition (e.g., \code{"12, 134, 147"}).
}

If \code{data_old} is provided, then the output will include:
\itemize{
\item The message (\code{text}) associated with the tested condition (e.g., \code{"Nb and list of IDs with invalid Age value"})
\item The number of rows meeting the condition in the dataset \code{data} (e.g., \code{"3"}), and
\item The list of unique identifiers meeting the given condition in the dataset \code{data}, and following the syntax: \if{html}{\out{<br>}}
\code{"[IDs present in both 'data_old' and 'data'] IDs present in 'data' only"}
}
}
\description{
Check a data frame (\code{data}) for a set of conditions (\code{cond}).
The function will count how many observations satisfy that particular condition,
and list the corresponding unique identifiers (from \code{varname}).
}
\examples{

# Example based on epiuf::DummyData (100 obs.)

# Check in the first 50 obs. of epiuf::DummyData the number and list of the IDs  
# of cases vaccinated with the brand Pfizer 
# - Output should be 4 patients 
dataCheck(data = epiuf::DummyData[1:50,], 
          cond = "CovVaccBr == 'Pfizer' & Case == 1", 
          text = "Cases vaccinated with Pfizer", 
          varname = "ID")

# Check same condition on the whole 100 obs. of epiuf::DummyData, but considering 
# that the first 50 obs. have already checked in a previous data checking session 
# - Output should be 8 patients, including 4 that have already been checked  
dataCheck(data = epiuf::DummyData[], 
          data_old = epiuf::DummyData[1:50,],
          cond = "CovVaccBr == 'Pfizer' & Case == 1", 
          text = "Cases vaccinated with Pfizer", 
          varname = "ID")

}
