% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDataSet.R
\name{datatoaddset}
\alias{datatoaddset}
\title{appendDataset}
\usage{
datatoaddset(maindata, datatoadd)
}
\arguments{
\item{maindata}{A dataset  for pooled data}

\item{datatoadd}{A dataset to Add to pooled}
}
\value{
The pooled dataset
}
\description{
appendDataset
}
