% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring.r
\name{deleteKeyring}
\alias{deleteKeyring}
\title{Delete  a Keyring}
\usage{
deleteKeyring(keyring_id)
}
\arguments{
\item{keyring_id}{A character string specifying the keyring name to delete}
}
\value{
A character string containing the password for further API calls.
}
\description{
This function checks if the specified keyring exists, and if so delete it.
}
\examples{
\dontrun{
# Set a secret for a fictional country called 'DB'
deleteKeyring("DB")
}
}
