% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring.r
\name{deleteKeyringAll}
\alias{deleteKeyringAll}
\title{Delete all specific keyrings}
\usage{
deleteKeyringAll()
}
\description{
This function deletes all keyrings that are identified by the \code{listKeyring} function.
It iterates over each keyring returned by \code{listKeyring} and deletes them.
}
\examples{
# To delete all keyrings identified by `listKeyring` function
deleteKeyringAll()
}
