% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{desc}
\usage{
desc(data, vars, labels = NULL)
}
\arguments{
\item{data}{name of the dataset}

\item{vars}{the variables list to describe (numeric or factor)}

\item{labels}{labels for the variables list}
}
\value{
a descriptive table
}
\description{
Aim: to describe numeric variables or categorical variables (must be in factor) in a table.
}
\examples{
\dontrun{
   table <- desc(data = df, vars = c("age","sex","fluvaccany"), 
   labels = c("Age", "Sex", "Seasonal influenza vaccination") )
}

}
