% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descBy.R
\name{describBy}
\alias{describBy}
\title{describBy}
\usage{
describBy(data, vars, by, labels = NULL)
}
\arguments{
\item{data}{name of the dataset}

\item{vars}{the variables list to describe (numeric or factor)}

\item{by}{variable to compare (must be a factor and have at least two categories)}

\item{labels}{labels for the variables list (var names used if NULL)}
}
\value{
a descriptive table
}
\description{
Aim: to describe numeric variables or categorical variables (must be in factor) according to another variable in a table.
}
\examples{
\dontrun{
   table <- descBy(data = df, vars = c("age","sex","fluvaccany"), 
   labels = c("Age", "Sex", "Seasonal influenza vaccination"), 
   by = "lab_flu"))
}

}
