% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_month_year_tools.R
\name{distanceIsoWeeks}
\alias{distanceIsoWeeks}
\title{distanceIsoWeeks}
\usage{
distanceIsoWeeks(date = date, origin = "2019-09-30")
}
\arguments{
\item{date}{A date Value}

\item{origin}{A reference date value}
}
\value{
the nimber of isoweek between the two dates
}
\description{
distanceIsoWeeks
}
\examples{
distanceIsoWeeks(Sys.Date(),"2000-01-01")
}
