% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRedCap.r
\name{downloadRedCap}
\alias{downloadRedCap}
\title{Download REDCap Data}
\usage{
downloadRedCap(password, country, filepath = NULL)
}
\arguments{
\item{password}{A character string representing the API token for the desired REDCap project.}

\item{country}{A character string representing the country/project}

\item{filepath}{A character string representing the desired file path in which the downloaded data will be saved.}
}
\value{
A CSV file containing the REDCap data.
}
\description{
This function downloads REDCap data from a specified project and saves it as a CSV file. If no file path has previously been set,
it defaults to the working directory of the user
}
\examples{
\dontrun{
# Please provide your own API token to use this example
# If you haven't already, you can set and retrieve password for REDCap using
# the createKeyring() or grabKeyring function. 
# Run this function or visit it's vignette in epiuf for more information.

# Begin the download process using another function
accessREDCap(country)
}
}
