% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{epitable}
\alias{epitable}
\title{Cross tabulation ( 2by2 table).}
\usage{
epitable(
  data,
  out,
  exp,
  epiorder = TRUE,
  missing = FALSE,
  row = FALSE,
  col = FALSE,
  fisher = TRUE
)
}
\arguments{
\item{data}{The dataframe to be analysed}

\item{out}{"Outcome" as numbers, factors or text}

\item{exp}{"Exposure" as numbers, factors or text. short syntax is available
see help(epifield)}

\item{epiorder}{Should data be reordered to respect epi tables , default to TRUE}

\item{missing}{Boolean if FALSE, missing are not included in the table.
A summary output of number of missing values is added at the end}

\item{row}{"Row percentages"}

\item{col}{"Col percentages"}

\item{fisher}{TRUE by default, display the fisher exact probability.
If table is larger than 2*2 then Fisher is not calculated}
}
\value{
An array containing  values of \code{
table : The resulting table
rowperc : The optional row percentage
colperc : The optional col percentage  
chisq : Chi Square value
p : Estimated probability of this distribution
fischer : Exact probaility
missing : Number of missing values

}
}
\description{
\code{epitable} Display a cross tabulation of two variables optionnaly with
row or col percentages. Chi Square with associated p.value are calculated.
If table contain binary variable, then epiorder function is apply on the two variable
to get a resulting table compatible with usual epidemiology interpretation.
0/1 variables are transformed into Yes/No and Yes is displayed before No
Exposed Cases appear on upper left part of the table.
}
\examples{
data <- data.frame(id = 1:10,
                   cases = c(rep(1,3), rep(0,7)),
                   vacc = sample(c(0,1), replace = TRUE, size = 10))
data[8,2]<-NA                    
table(data$cases, data$vacc, useNA = "always")
result <- epitable(data,cases,vacc)
epitable(data,cases,vacc,epiorder=FALSE)

epitable(data,"cases","vacc") 
epitable(data,out=cases,exp=vacc,missing=TRUE) 
epitable(data,out=cases,exp=vacc,row=TRUE) 

table <- result
result$table

data <- data.frame(id = 1:10,
                   cases = c(rep(1,3), rep(0,7)),
                   vacc = sample(c(0,1,2), replace = TRUE, size = 10))

}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\seealso{
\code{\link{freq}} for frequency distributions
}
\author{
Gilles Desve
}
