% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring.r
\name{epiufKeyring}
\alias{epiufKeyring}
\title{Generate a standardized keyring ID}
\usage{
epiufKeyring(keyring_id)
}
\arguments{
\item{keyring_id}{A string specifying the base keyring ID.}
}
\value{
A string representing the standardized keyring ID.
}
\description{
This function appends "_epiufkeyring" to the given keyring ID to generate a standardized keyring ID.
It's useful for maintaining consistency in naming keyrings within your system.
}
