% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring.r
\name{existsKeyring}
\alias{existsKeyring}
\title{Check if a specific keyring exists}
\usage{
existsKeyring(keyring_id)
}
\arguments{
\item{keyring_id}{A string specifying the base keyring ID.}
}
\value{
A logical value; \code{TRUE} if the keyring exists, otherwise \code{FALSE}.
}
\description{
This function checks for the existence of a keyring. It first standardizes the keyring ID
using \code{epiufKeyring} function and then checks if this standardized ID is present in the list of keyrings.
}
\examples{
# Check if a keyring exists
keyring_exists <- existsKeyring("myKeyring")
print(keyring_exists) # Output: TRUE or FALSE, depending on existence
}
