% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandMacro}
\alias{expandMacro}
\title{expandMacro}
\usage{
expandMacro(x)
}
\arguments{
\item{x}{A string containing a macro}
}
\value{
The string with macro processed and repalced by value
}
\description{
expandMacro
}
\examples{
varMacro <- "few word"
expandMacro("text with ${varMacro} inside")

}
