% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandVar.R
\name{expandVar}
\alias{expandVar}
\title{expandVar}
\usage{
expandVar(data, varname, valueslist)
}
\arguments{
\item{data}{A dataset which contains the variable with more than one value (data.frame).}

\item{varname}{The name of the variable.}

\item{valueslist}{List of expected values with names of the variables to be created.}
}
\value{
A dataset with the new columns added.
}
\description{
To expand a variable containing more than one value into new individual variables with Y/N values.
The list of expected values must be passed as parameters together with the name of the new variable to be created.
}
\examples{

# Create an example data frame 
data <- data.frame(Id = 1:4 , 
                    Vaccs = c("pfizer,moderna"," ", "pfizer", "moderna"))

# Specify the name of the new column and what value the new column represents 
brand <- list("pfizer"="pfizer",
             "moderna"="moderna"
              )
# Use the function
data <-  expandVar(data = data, varname = Vaccs, valueslist = brand)

# View the changed dataset 
print(data)
}
