% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandVarAll.R
\name{expandVarAll}
\alias{expandVarAll}
\title{expandVarAll}
\usage{
expandVarAll(data, action = NULL)
}
\arguments{
\item{data}{dataframe to be processed, the dico must be loaded}

\item{action}{optional specific dictionary for action}
}
\value{
data.frame
}
\description{
expandVarAll
}
