% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasourcestools.R
\name{exportSourceList}
\alias{exportSourceList}
\title{exportSourceList}
\usage{
exportSourceList(filename)
}
\arguments{
\item{filename}{A filename where to save the source list}
}
\value{
nothing
}
\description{
exportSourceList
}
\examples{
# nothing to show 
}
