% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{externalFile}
\alias{externalFile}
\title{Locate External File Across Different Paths}
\usage{
externalFile(extfile)
}
\arguments{
\item{extfile}{Character. The name of the external file to locate.}
}
\value{
Character. The path of the located external file.
}
\description{
This function takes an external file name and searches for its existence across different possible locations.
It first checks the current working directory, then the package's "extdata" directory, and finally a relative "inst/extdata" path.
Raises an error if the file cannot be found in any of the locations.
}
\examples{
\dontrun{
externalFile("data.csv")  # Searches for "data.csv" in various locations
}

}
\seealso{
\code{\link{file.exists}}, \code{\link{system.file}}
}
