% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorUsing.R
\name{factorUsing}
\alias{factorUsing}
\title{Title}
\usage{
factorUsing(data, varname, dicoCode, dicoLabel)
}
\arguments{
\item{data}{}

\item{varname}{}

\item{dicoCode}{}

\item{dicoLabel}{}
}
\value{

}
\description{
Title
}
