% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorUsingDico.R
\name{factorUsingDico}
\alias{factorUsingDico}
\title{factorUsingDico}
\usage{
factorUsingDico(data, dictionary = NULL, dicos = NULL)
}
\arguments{
\item{data}{The dataset to process (data.frame).}

\item{dictionary}{The dictionary (epiuf structure).}

\item{dicos}{The dicos (epiuf structure).}
}
\value{
Data.frame
}
\description{
This function aims to factorise and label coded variables based off the data dictionary.
}
