% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{fileExt}
\alias{fileExt}
\title{Extract File Extension From File Path or URL}
\usage{
fileExt(text)
}
\arguments{
\item{text}{Character. The full path or URL of the file.}
}
\value{
Character. The extension of the file. Returns an empty string if no extension is found.
}
\description{
This function takes a file path or URL and returns the file extension.
Unlike functions like \code{tools::file_ext}, this function handles both paths and URLs.
}
\examples{
fileExt("path/to/file.txt")  
fileExt("https://example.com/file.csv") 
fileExt("no_extension_file") 

}
