% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{fileFindReplace}
\alias{fileFindReplace}
\title{Find and Replace Text in a File}
\usage{
fileFindReplace(
  filename,
  pattern,
  replacement,
  wholeword = TRUE,
  ignore.case = FALSE,
  listonly = FALSE
)
}
\arguments{
\item{filename}{Character. The name of the file to modify.}

\item{pattern}{Character vector. A vector of patterns to search for in the file.}

\item{replacement}{Character vector. A vector of replacements for the patterns found.}

\item{wholeword}{Logical. If TRUE, only whole words will be replaced. Defaults to TRUE.}

\item{ignore.case}{Logical. If TRUE, the function performs a case-insensitive search. Defaults to FALSE.}

\item{listonly}{Logical. If TRUE, instead of modifying the files, a summary of the changes that would be made
is displayed.. This is used to check changes before applying them. Defaults to FALSE.}
}
\value{
NULL. The function modifies the file in-place.
}
\description{
This function reads a file and replaces specified patterns with given replacements.
The function can work with multiple patterns and replacements, and can also perform
case-sensitive or case-insensitive searches.
}
\examples{
\dontrun{
fileFindReplace("example.txt", "oldText", "newText")
fileFindReplace("example.txt", c("oldText1", "oldText2"), "newText")
}
}
\seealso{
\code{\link{txtFindReplace}}, \code{\link{xlsxFindReplace}}
}
