% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{fillimage}
\alias{fillimage}
\title{fillimage}
\usage{
fillimage(
  onesheet,
  image,
  line,
  col,
  wide = 7,
  high = 4,
  unit = "in",
  spec.dpi = 300,
  wb = NULL
)
}
\arguments{
\item{onesheet}{}

\item{image}{}

\item{line}{}

\item{col}{}

\item{wide}{}

\item{high}{}

\item{unit}{}

\item{spec.dpi}{}

\item{wb}{}
}
\description{
fillimage
}
\examples{
fillimage(onesheet = cells,image = graphname,line=2, col = 2, wb=template)
}
