% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDrive.R
\name{findDrive}
\alias{findDrive}
\title{findDrive}
\usage{
findDrive(message = FALSE)
}
\arguments{
\item{message}{boolean to print information messages or not}
}
\value{
path to the shared drive
}
\description{
Quick way to find the path to the root Shared Drive (e.g., "G:/Shared drive").
It will identify the letter associated to the Drive (e.g., D:, G:, etc)
and then the name of the root folder (i.e., "Shared drive" or "Drive partagé").
Indeed, depending on the user language, it could be labelled in english or French.
Only these tow languages are implemented so far.
}
\examples{
setPath("DRIVE", findDrive())
getPath("DRIVE")

}
\author{
STRAP team \email{strap@epiconcept.fr}
}
