% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixencoding.R
\name{fixEncoding}
\alias{fixEncoding}
\title{fixEncoding}
\usage{
fixEncoding(df, originalEncoding = "UTF-8")
}
\arguments{
\item{df}{A dataset to encode with new character set}

\item{originalEncoding}{Either latin1 or UTF-8 which are the two encoding avalaible in R if}
}
\value{
The encoded dataset
}
\description{
Add the encoding value to character vectors which contain special characters
(including accents)
regular ascci values stay as "unknown"
}
