% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{formatCells}
\alias{formatCells}
\title{formatCells}
\usage{
formatCells(onesheet, line, col, style = NULL)
}
\arguments{
\item{onesheet}{A sheet object from xlsx package}

\item{line}{The line where to paste value}

\item{col}{The col where to paste value}

\item{style}{A style created by createXlsxStyle}
}
\value{
nothing
}
\description{
formatCells
}
