% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{freq}
\usage{
freq(x, y = NULL, missing = FALSE, quietly = FALSE)
}
\arguments{
\item{x}{As numbers, factors or text.}

\item{y}{As numbers, factors or text.}

\item{missing}{If FALSE, then missing values are not included in the table. A summary output of number of missing values is added at the end.}

\item{quietly}{No output, only return value}
}
\value{
An array containing values.
}
\description{
This function is used to display a frequency distribution.
}
\examples{

# Create an example data frame
set.seed(42)

df <- data.frame(
Gender = sample(c("Male", "Female"), 30, replace = TRUE),
Department = sample(c("HR", "Finance", "IT", "Marketing"), 30, replace = TRUE),
Satisfaction = sample(c("Low", "Medium", "High"), 30, replace = TRUE)
)

# Create frequency distribution table 
freq(df$Gender)



}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\seealso{
\code{\link{table}} for 2by2 tables
}
\author{
Gilles Desve
}
