% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getActionGroup}
\alias{getActionGroup}
\title{getActionGroup}
\usage{
getActionGroup(actiontag)
}
\arguments{
\item{actiontag}{Name of the action group to retrieve (character)}
}
\value{
Dataset of dictionary actions containing all the variables with actions of type action group
}
\description{
Return the dictionary action (see \code{getDictionaryActions()}) filtered
for a given \code{action_group}.
}
\seealso{
\code{\link[=getDictionaryActions]{getDictionaryActions()}}, \code{\link[=getVarAction]{getVarAction()}}
}
