% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{getAllFilepaths}
\alias{getAllFilepaths}
\title{Get list of all filepaths and subsidiaries within a directory}
\usage{
getAllFilepaths(directory, filetypes = NULL, ignore.these = NULL)
}
\arguments{
\item{directory}{The filepath where you want to search.}

\item{filetypes}{The types of file you want to list (optional) (input is eg. c("R", "Rmd").}

\item{ignore.these}{The names of any folders you wish to ignore.}
}
\value{
Returns list of complete filepaths
}
\description{
This function starts with a given folder and returns the filepaths of all files
in all the folders. You can specify specific folders to ignore, and specific filetypes
to find.
}
\examples{
\dontrun{
allpaths <- getAllFilepaths(getwd(),filetypes = c("R","Rmd","txt"), ignore.these = c(".git", ".gitignore",".Rproj.user"))
}
}
