% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDelayIf.R
\name{getDelayIf}
\alias{getDelayIf}
\title{getDelayIf}
\usage{
getDelayIf(data, FirstDateName, SecondDateName, ...)
}
\arguments{
\item{data}{The dataset containing the values (data.frame).}

\item{FirstDateName}{The first date (as.Date).}

\item{SecondDateName}{The second date (as.Date).}

\item{...}{One or more logical condition to validate. If one of the condition is not TRUE,
NA is returned instead of the calculated delay.}
}
\value{
A vector with the same length of the input dataset.
}
\description{
This function returns the delay between two dates with a condition.
If the condition is validated then the delay in days from date one to date two is returned,
else NA is returned.
}
\examples{

# Create a random data frame with two date variables
df <- data.frame(date1 = c("2023-01-01", "2023-01-02", "2023-01-03", "2023-01-04")
,date2 = c("2023-02-01", "2023-02-02", "2023-04-03", "2023-03-04")
,id = c("a", "b", "c", "d"))

# Convert them to the correct format 
df$date1 <- as.Date(df$date1)
df$date2 <- as.Date(df$date2)

# Use the function 
df$test <- ifelse(getDelayIf(df, date1, date2)>40, 1, 0) 

}
