% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maindialog.R
\name{getDialog}
\alias{getDialog}
\title{getDialog  get a value from the dialog system}
\usage{
getDialog(value = NULL)
}
\arguments{
\item{value}{The name of the value to retrieve
If NULL then the dialog env is returned}
}
\value{
either a value or an Env
}
\description{
getDialog  get a value from the dialog system
}
