% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getDico}
\alias{getDico}
\title{getDico}
\usage{
getDico(diconame)
}
\arguments{
\item{diconame}{The name of one dico from the dicos structure}
}
\value{
A data set containing one dico (list of code/labels)
}
\description{
This function returns a data set containing one dico (list of code/labels).
}
