% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getDicos}
\alias{getDicos}
\title{getDicos}
\usage{
getDicos()
}
\value{
The dataset containing all the dicos in 3 columns: dico_name, label, code.
}
\description{
This function returns a dataset containing all stored dicos, including
the list of codes and labels, as returned by \code{getNewDictionaryLine("dicos")}.
}
\examples{
getDicos() 


}
\seealso{
\code{\link[=setDicos]{setDicos()}}
}
