% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getDictionaryActions}
\alias{getDictionaryActions}
\title{getDictionaryActions}
\usage{
getDictionaryActions()
}
\value{
A dataset of actions
}
\description{
Return the dictionary actions dataset stored in the environment
as returned by \code{getNewDictionaryLine("actions")}..
}
\examples{
getDictionaryActions()

 
}
\seealso{
\code{\link[=setDictionaryActions]{setDictionaryActions()}}, \code{\link[=getNewDictionaryLine]{getNewDictionaryLine()}}
}
