% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getDictionaryValue}
\alias{getDictionaryValue}
\title{getDictionaryValue}
\usage{
getDictionaryValue(varname, valuename = c("type", "dico", "unknowns"))
}
\arguments{
\item{varname}{The varname for which we will retrieve content of one column from the dictionary (character).}

\item{valuename}{Name of the column to retrieve from dictionary (character).}
}
\value{
A single value
}
\description{
Retrieve the value of one parameter's (column) in the dictionary, searching for the 'generic_name'.
Usual column values to retrieve are: type, dico and unknowns.
An error will occur if the column name is incorrect.
Return NA if searched varname is not found.
}
\examples{

# Create example dataset 
dic <- data.frame(generic_name = c("id", "idcountry", "age", "agegroup3", "agroup10"),
                  source_name = c("id", "idcountry", "age", "agegroup3", "agroup10"),
                  type = c("character", "character", "numeric", "factor", "factor"),
                  dico = c("","","","agegp3","agegp10"))

# Set the dictionary 
setDictionary(dictionary = dic)

# Use the function
getDictionaryValue("agegroup3","dico")

}
\seealso{
\code{\link[=getNewDictionaryLine]{getNewDictionaryLine()}}
}
