% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiOption.R
\name{getEpiOption}
\alias{getEpiOption}
\title{getEpiOption}
\usage{
getEpiOption(EpiOption)
}
\arguments{
\item{EpiOption}{name of the option to retrieve}
}
\value{
option value
}
\description{
retrieve an epiutils package option or parameter
}
\examples{
getEpiOption("optionname")


}
