% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasourcestools.R
\name{getFileList}
\alias{getFileList}
\title{getFileList  Return a list of files for specific status}
\usage{
getFileList(
  country = NULL,
  status = NULL,
  periodstart = NULL,
  periodend = NULL,
  sourcetype = NULL
)
}
\arguments{
\item{country}{The country code}

\item{status}{The current status searched for  (where CurrentStatus is TRUE )}

\item{periodstart}{Optional}

\item{periodend}{Optional}

\item{sourcetype}{Optional}
}
\value{
a dataset containing files informations
}
\description{
getFileList  Return a list of files for specific status
}
\examples{
\dontrun{  getFileList("FR","IMPORTED")} 
}
