% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMath.R
\name{getMean}
\alias{getMean}
\title{getMean}
\usage{
getMean(...)
}
\arguments{
\item{...}{further arguments passed to \code{base::mean()} or from other methods}

\item{x}{numeric, logical, date vector}
}
\value{
single numeric, logical, date mean value
}
\description{
Get the arithmetic mean value from a vector skipping NA values. If all values are NA, returns NA.
}
\examples{
# Numerics
getMean(c(1, 9))
getMean(c(1, 9, NA))
getMean(c(as.Date("2023-01-01"), 
         NA, 
         as.Date("2023-01-31")))


# If missing values
## using basic mean() function from base R
mean(c(1:9, NA)) #NA value takes over the other values by default
mean(c(1:9, NA), na.rm = TRUE) #Need to add na.rm = TRUE to get the min value despite NAs
mean(c(NA, NA), na.rm = TRUE) #If NAs only, returns NaN
## using getMean, we do not have these issues
getMean(c(NA, 1:9))
getMean(c(NA, NA))

}
\seealso{
\code{\link[base:mean]{base::mean()}} and \code{\link[=getMin]{getMin()}}, \code{\link[=getMax]{getMax()}},
\code{\link[=getMedian]{getMedian()}} from \code{epiuf} package
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
