% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMath.R
\name{getMedian}
\alias{getMedian}
\title{getMedian}
\usage{
getMedian(...)
}
\arguments{
\item{...}{numeric, logical, date vector}
}
\value{
single numeric, logical, date median value
}
\description{
Get the median value from a vector skipping NA values. If all values are NA, returns NA.
}
\examples{
# Numerics
getMedian(c(1, 9))
getMedian(c(1, 9, NA))
getMedian(c(as.Date("2023-01-01"), 
         NA, 
         as.Date("2023-01-31")))


# If missing values
## using basic mean() function from base R
median(c(1:9, NA)) #NA value takes over the other values by default
median(c(1:9, NA), na.rm = TRUE) #Need to add na.rm = TRUE to get the min value despite NAs
median(c(NA, NA), na.rm = TRUE)
## using getMedian, we do not have this issue
getMedian(c(NA, 1:9))
getMedian(c(NA, NA))

}
\seealso{
\code{\link[stats:median]{stats::median()}} and \code{\link[=getMean]{getMean()}}, \code{\link[=getMin]{getMin()}}, \code{\link[=getMax]{getMax()}} from \code{epiuf} package
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
