% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMath.R
\name{getMin}
\alias{getMin}
\title{getMin}
\usage{
getMin(...)
}
\arguments{
\item{...}{Numeric or character vector arguments.}
}
\value{
A single numeric or character element.
}
\description{
Get the minimum value from a vector skipping NA values. If all values are NA, returns NA. Can be letter, number, date etc.
}
\examples{
# Numerics
getMin(c(1, 9))
getMin(1, 9, NA) #c() is not mandatory
getMin(as.Date("2023-01-01"), 
       NA, 
       as.Date("2023-01-31"))

# Characters
getMin(LETTERS)
## Note: For characters, getMin() will return the first month 
## according to alphabetical order (i.e., "April") 
getMin(month.name)
getMin(NA, month.name)

# If missing values
## using basic min() function from base R
min(c(1:9, NA)) #NA value takes over the other values by default
min(c(1:9, NA), na.rm = TRUE) #Need to add na.rm = TRUE to get the min value despite NAs
min(c(NA, NA), na.rm = TRUE) #If NAs only, returns Inf and a warning
## using getMin, we do not have these issues
getMin(c(NA, 1:9))
getMin(c(NA, NA))

}
\seealso{
\code{\link[base:Extremes]{base::min()}} and \code{\link[=getMax]{getMax()}}, \code{\link[=getMean]{getMean()}},
\code{\link[=getMedian]{getMedian()}} from \code{epiuf} package
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
