% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getNewDictionaryLine}
\alias{getNewDictionaryLine}
\title{getNewDictionaryLine}
\usage{
getNewDictionaryLine(mode = "dictionary")
}
\arguments{
\item{mode}{Type of line to return (character). Could be dictionary, dicos, or actions
(one of the three tabs of the Excel dictionary file, see \code{openDictionary()})}
}
\value{
An empty record in a data frame of type \code{mode}
}
\description{
getNewDictionaryLine
}
\examples{
getNewDictionaryLine(mode = "dictionary")
getNewDictionaryLine(mode = "dicos")
getNewDictionaryLine(mode = "actions")


}
\seealso{
\code{\link[=openDictionary]{openDictionary()}}
}
