% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSnippets.R
\name{getOS}
\alias{getOS}
\title{Project Name : epiuf
Script Name  : snippets.R
GitHub repo  : STRAP-epiuf
Summary      :
Date created : 2023-06-01
Author       : Lore Merdrignac
Date reviewed:
Reviewed by  :
Description -------------------------------------------------------------
Borrowed and adapted from
https://stackoverflow.com/questions/62220596/can-r-packages-add-code-snippets-to-users-snippet-files
Changes Log -------------------------------------------------------------}
\usage{
getOS()
}
\value{
character string naming the operator system in use
}
\description{
getOS
}
\details{
Returns the user OS in lower case
}
\examples{
getOS()

}
\seealso{
\code{\link[=Sys.info]{Sys.info()}}
}
\author{
Lore Merdrignac
}
