% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{getPath}
\alias{getPath}
\title{getPath}
\usage{
getPath(pathname)
}
\arguments{
\item{pathname}{Path label specified in "".}
}
\value{
The path saved under pathname label.
}
\description{
This function retrieves a named Path previously defined in 'setPath()'.
}
\examples{

\dontrun{
# First set the path directory and give it a label 
setPath(pathname = "SOURCES", 
    path = "c:/dev/Resources", 
    makedir = "Never")

# Get the file path using the function 
getPath("SOURCES")
}

}
\seealso{
\code{\link[=setPath]{setPath()}}
}
