% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefSite.R
\name{getRefSiteLabel}
\alias{getRefSiteLabel}
\title{Get the Label corresponding to the site ID}
\usage{
getRefSiteLabel(ID)
}
\arguments{
\item{ID}{Character string to search in the \code{"ID"} column from \code{epiuf::RefSite} reference table}
}
\value{
Character string or a data frame if multiple answers
}
\description{
Get the Label corresponding to the ID code of the country/study site from \code{epiuf::RefSite} reference table
}
\examples{
\dontrun{
getRefSiteLabel()
}
getRefSiteLabel(ID = "SI")
getRefSiteLabel(ID = "UK")
\dontrun{
getRefSiteLabel(ID = "es")
}

}
\seealso{
\itemize{
\item \code{\link{getRefSiteLocName}}
\item \code{\link{getRefSiteID}}
}
For more details see the corresponding vignette: \code{vignette("RefSite")}
}
\author{
Lore Merdrignac \email{l.merdrignac@epiconcept.fr}
}
