% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefSite.R
\name{getRefSiteLocName}
\alias{getRefSiteLocName}
\title{Get the LocationName corresponding to the site ID}
\usage{
getRefSiteLocName(ID)
}
\arguments{
\item{ID}{Character string to search in the \code{"ID"} column from \code{epiuf::RefSite} reference table}
}
\value{
Character string or a data frame if multiple answers
}
\description{
Get the LocationName corresponding to the ID code of the country/study site from \code{epiuf::RefSite} reference table
}
\examples{
\dontrun{
getRefSiteLocName()
}
getRefSiteLocName(ID = "SI")
getRefSiteLocName(ID = "UK")
\dontrun{
getRefSiteLocName(ID = "es")
}

}
\seealso{
\itemize{
\item \code{\link{getRefSiteLabel}}
\item \code{\link{getRefSiteID}}
}
For more details see the corresponding vignette: \code{vignette("RefSite")}
}
\author{
Lore Merdrignac \email{l.merdrignac@epiconcept.fr}
}
