% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasourcestools.R
\name{getSourceList}
\alias{getSourceList}
\title{getSourceList}
\usage{
getSourceList()
}
\value{
the dataset containing all source information
}
\description{
getSourceList
}
\examples{
\dontrun{ getSourceList() }
}
