% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getWord}
\alias{getWord}
\title{split a character string into word and return word x}
\usage{
getWord(tosearch, item = 1, pattern = "\\\\W+")
}
\arguments{
\item{tosearch}{The character string to be searched for}

\item{item}{The \code{item} word in the string}

\item{pattern}{The sep as a regular expression pattern, default to word separator}
}
\value{
the word at \code{item} position in tosearch
}
\description{
split a character string into word and return word x
}
\examples{
getWord("aaa  bb,cc.",3)
getWord("aaa-bb-cc",2,pattern="-")
}
