% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{getWorkbook}
\alias{getWorkbook}
\title{getWorkbook}
\usage{
getWorkbook()
}
\value{
The current workbook in memory
}
\description{
Retrieve the current workbook saved in the environment.
}
\examples{
getWorkbook()

}
