% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{globalFileReplace}
\alias{globalFileReplace}
\title{Batch Find and Replace in R Scripts within a Directory}
\usage{
globalFileReplace(directory, oldtext, newtext, wholeword, ignore.case = FALSE)
}
\arguments{
\item{directory}{The directory where R scripts are located.}

\item{oldtext}{The text to search for in R scripts.}

\item{newtext}{The text that will replace \code{oldtext}.}

\item{wholeword}{Logical, if TRUE only whole words will be replaced.}

\item{ignore.case}{Logical, if TRUE the function ignores case.}
}
\value{
No return value; the function modifies the files in-place.
}
\description{
This function finds and replaces specified text in all R scripts (.R or .r) within a given directory.
}
\examples{
\dontrun{
globalFileReplace("./myRScripts/", "oldFunc", "newFunc")
}
}
