% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring.r
\name{grabKeyring}
\alias{grabKeyring}
\title{Retrieve password from a keyring}
\usage{
grabKeyring(keyring_id)
}
\arguments{
\item{keyring_id}{A string specifying the keyring ID.}
}
\value{
A string with the keyring password or an empty string if keyring is not found.
}
\description{
This function retrieves the password from a specified keyring.
If the keyring does not exist, it issues a warning.
}
