% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSnippets.R
\name{importSnippets}
\alias{importSnippets}
\title{Import snippets}
\usage{
importSnippets()
}
\value{
boolean \code{invisible(FALSE)} if nothing was added, \code{invisible(TRUE)} if snipped definitions were added
}
\description{
\code{addPackageSnippets} copies all missing snippet definitions
from 'inst/rstudio/Rsnippets.txt' to the RStudios user snippet location
}
\examples{
\dontrun{importSnippets()}

}
