% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{isDate}
\alias{isDate}
\title{isDate}
\usage{
isDate(date)
}
\arguments{
\item{date}{date object or variable that wish to test}
}
\value{
logical
}
\description{
Uses the base function class to asses if an object is a date or not
}
\examples{
isDate(epiuf::DummyData$EnrolmentDate)
}
\seealso{
For more details see the link below to access the vignette:
\href{../doc/epiuf_package.html}{\code{vignette("epiuf_package")}}
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
