% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{isoYearWeek}
\alias{isoYearWeek}
\title{isoYearWeek}
\usage{
isoYearWeek(date, weekformat = NULL)
}
\arguments{
\item{date}{A date Value (as.date or Sys.date). The date is specified in "".}

\item{weekformat}{A string to represent the week separator. Default is "w".
It can be changed to be one or more characters (eg : "-w").}
}
\value{
The ISO year and week of the Date in YYYYwXX format (2020w03).
}
\description{
Returns the ISO year and week in character string format (e.g., 2020w03) of a date object.
}
\examples{

# Extract the ISO year and week of the current date
isoYearWeek(Sys.Date())

# Extract the ISO year and week of the specified date
isoYearWeek(as.Date("2020-12-31"))

# Separate the year and week in the output by specifying the 'weekformat' 
isoYearWeek(as.Date("2020-12-31"), weekformat="-w")

}
\seealso{
\code{\link[=isoYear]{isoYear()}}, \code{\link[=isoWeek]{isoWeek()}}, \code{\link[=lpad]{lpad()}} and \code{\link[base:format]{base::format()}}
}
