% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isVar}
\alias{isVar}
\title{isVar fonction WIP do not use
exists look only in GlobalEnv and parent, is.var will search from current and parent until global but not in base}
\usage{
isVar(what = "")
}
\arguments{
\item{what}{An object name to find}
}
\value{
TRUE if found
}
\description{
isVar fonction WIP do not use
exists look only in GlobalEnv and parent, is.var will search from current and parent until global but not in base
}
\examples{
isVar(test)

}
