% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiOption.R
\name{listEpiOption}
\alias{listEpiOption}
\title{listEpiOption}
\usage{
listEpiOption(pattern = "*")
}
\arguments{
\item{pattern}{A pattern for options to be listed}
}
\value{
List of options defined with setEpiOption
}
\description{
listEpiOption
}
\examples{

 listEpiOption()

}
