% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring.r
\name{listKeyring}
\alias{listKeyring}
\title{List specific keyrings}
\usage{
listKeyring(pattern = NULL)
}
\arguments{
\item{pattern}{A pattern to filter the list}
}
\value{
A character vector of keyring names containing "epiufkeyring".
}
\description{
This function lists all keyrings that contain the term pattern (default to "epiufkeyring") in their names.
It uses a regular expression to find matches and returns the filtered list.
}
\examples{
# Assuming you have keyrings named 'epiufkeyring_1', 'epiufkeyring_2', etc.
listed_keyrings <- listKeyring()
print(listed_keyrings)
}
