% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listVar}
\alias{listVar}
\title{listVar}
\usage{
listVar(dataset, pattern, regex = FALSE)
}
\arguments{
\item{dataset}{A dataset to explore}

\item{pattern}{Pattern representing varname}

\item{regex}{Should the pattern be used as regex expression or use classical "joker" ? and *}
}
\value{
list of the variables matching the pattern
}
\description{
listVar
}
\examples{
data <- data.frame(Id = 1:4 ,  
                   vaccage = c(34,45, 50,22 ),
                   symp = c("Y","Y","N","N"),
                   vaccboost=c("N","Y","N","Y"))
listVar(data,"symp")
listVar(data,"vac*")                    
                   

}
