% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libutils.R
\name{loadLib}
\alias{loadLib}
\title{loadLib}
\usage{
loadLib(libname)
}
\arguments{
\item{libname}{Name of the package you want to install and load.}
}
\value{
The library name will print in the console if loaded, else "false".
}
\description{
This function loads a library package and will ask for confirmation before loading.
}
\examples{

# Load the package 'utils' 
loadLib("utils")

}
\seealso{
For more details see the help vignette: \cr
\code{vignette("epiuf_package", package = "epiuf")} \cr
\href{"../doc/epiuf_package.html"}{epiuf_package}
}
