% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg.R
\name{logreg}
\alias{logreg}
\title{logreg}
\usage{
logreg(outvar, expvar, data, full = TRUE, quietly = FALSE)
}
\arguments{
\item{outvar}{As numbers, factors or text. Represent the outcome}

\item{expvar}{As character : model to be tested as var1 + var2 + var3}

\item{data}{The dataframe containg vars}

\item{full}{For full output}

\item{quietly}{If TRUE, supress direct output}
}
\value{
An array containing  resulting summary
#' @examples
logreg()
}
\description{
\code{logreg} Do a logistic regression using submitted model .
#'
}
\references{
Based on: \emph{Epi6} and \emph{Stata} functionnality,
available at \url{https://github.com/}.
}
\author{
Gilles Desve
}
