% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maindialog.R
\name{maindialog}
\alias{maindialog}
\title{maindialog Run a dialog to drive the project}
\usage{
maindialog()
}
\value{
Run the application (return a shiny R object to print() )
}
\description{
maindialog Run a dialog to drive the project
}
