% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDico.R
\name{matchDico}
\alias{matchDico}
\title{Title}
\usage{
matchDico(data, mydicos, varname, diconame)
}
\arguments{
\item{data}{}

\item{mydicos}{}

\item{varname}{}

\item{diconame}{}
}
\value{

}
\description{
Title
}
