% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{Month}
\alias{Month}
\title{Month}
\usage{
Month(date)
}
\arguments{
\item{date}{A date value either Sys.Date or as.Date.}
}
\value{
The month of the date in numeric.
}
\description{
Returns the month of a date object in numeric format.
}
\examples{

# Get the month of today's date 
Month(Sys.Date())

# Get the month of the specified date 
Month(as.Date("2024/05/23"))

}
\seealso{
\code{\link[=abrvMonth]{abrvMonth()}} and \code{\link[base:format]{base::format()}}
}
