% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_drop.R
\name{moyenne}
\alias{moyenne}
\title{moyenne d’un vecteur >> Obsolete replaced by getMean}
\usage{
moyenne(x)
}
\arguments{
\item{x}{un vecteur numerique}
}
\value{
la fonction renvoie la moyenne d'un vecteur
}
\description{
Une fonction pour faire une moyenne en enlevant les valeurs manquantes
}
