% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{openDictionary}
\alias{openDictionary}
\title{openDictionary}
\usage{
openDictionary(filename)
}
\arguments{
\item{filename}{Character string, path to the file (xls) containing a dictionary. The dictionary will be loaded.}
}
\value{
Nothing
}
\description{
Imports a dictionary into the environment.
}
\examples{
# example code
 
 file1 <- externalFile("genericdictionary.xlsx")
 dictionary <- openDictionary(file1)

}
