% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasourcestools.R
\name{openSourceList}
\alias{openSourceList}
\title{openSourceList}
\usage{
openSourceList(filename = pathToFile("SOURCES", "datasources.xls"))
}
\arguments{
\item{filename}{Name of the excel file containing sources information}
}
\value{
nothing
}
\description{
openSourceList
}
\examples{
\dontrun{ openSourceList("datasources.xls")}
}
