% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{openXlsx}
\alias{openXlsx}
\title{openXlsx}
\usage{
openXlsx(filename = "")
}
\arguments{
\item{filename}{Name of the Excel file specified in "" (xlsx). Or the name of the filepath (see example).}
}
\value{
The wb (which is also saved as internal variable).
}
\description{
This function loads an excel workbook into the environment.
}
\examples{

# Filepath created for the workbook 
xls_file <- externalFile("excelfile.xlsx")

# Load workbook into the environment 
wb <- openXlsx(xls_file)

}
