% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{pathToFile}
\alias{pathToFile}
\title{pathToFile}
\usage{
pathToFile(pathname, filename)
}
\arguments{
\item{pathname}{Label for saved path}

\item{filename}{Name of the file to retrieve, may contain subdir}
}
\value{
the full name of the file
}
\description{
Constructs an absolute path to a file (e.g a folder) using a previously set saved path name (see setPath) and the file name.
}
\examples{
pathToFile("data","cleanedrecords.csv")


}
