% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pos}
\alias{pos}
\title{pos}
\usage{
pos(pattern, stosearch)
}
\arguments{
\item{pattern}{A pattern to search in stosearch}

\item{stosearch}{A character string}
}
\value{
Position of pattern in stosearch
}
\description{
pos
}
\examples{
pos("/","test/string")

}
