% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDataCheck.R
\name{printDataCheck}
\alias{printDataCheck}
\title{Title}
\usage{
printDataCheck(
  data,
  data_old = NULL,
  parameters,
  header = "",
  threshold = NULL,
  varname = "id"
)
}
\arguments{
\item{data}{}

\item{data_old}{}

\item{parameters}{}

\item{header}{}

\item{threshold}{}

\item{varname}{}
}
\value{
data.frame
}
\description{
Title
}
\examples{
printDataCheck(data=df, data_old=NULL,  parameters = parameters, header="Essential Data Checks", threshold=NULL , varname="idCheck")
}
