% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printidcond.R
\name{PrintIDCond}
\alias{PrintIDCond}
\title{PrintIDCond}
\usage{
PrintIDCond(data, threshold, cond, column = "IdData")
}
\arguments{
\item{data}{A dataframe to look for cond}

\item{threshold}{Max number of ID to print}

\item{cond}{A condition}

\item{column}{Column nameto print}
}
\value{
nothing but print using cat
}
\description{
PrintIDCond
}
\examples{
data <- data.frame(Id = 1:4 , 
                    Vaccs = c("pfizer"," ", "pfizer", "moderna"))
PrintIDCond(df,"Vaccs=='pfizer'",20,"Id")
                     
}
