% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCheck.R
\name{printIf}
\alias{printIf}
\title{printIf}
\usage{
printIf(data, cond, text = "", threshold = NULL, varname = "id", na.rm = FALSE)
}
\arguments{
\item{data}{A dataframe to look for the condition (data.frame).}

\item{cond}{The condition to check.}

\item{text}{The message to print (if empty the condition is used).}

\item{threshold}{Cutoff number of outputs meeting this condition.}

\item{varname}{Column name of output to print.}

\item{na.rm}{Remove missing output values in the list to print, by default is False.}
}
\value{
Message to print as list
}
\description{
Prints specified text when a certain condition is met.
}
\examples{
df <- data.frame(ID = 1:4, 
                 Vaccs = c("pfizer"," ", "pfizer", "moderna"))
printIf(data = df, 
        cond = Vaccs == "pfizer", 
        threshold = 30, 
        text = "Pfizer vaccine", 
        varname = "ID")
        
}
