% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCheck.R
\name{printIf2}
\alias{printIf2}
\title{printIf2
prints the list of IDs meeting the specified condition}
\usage{
printIf2(
  data,
  cond,
  text = "",
  threshold = NULL,
  varname = "id",
  na.rm = FALSE,
  data_old = NULL
)
}
\arguments{
\item{data}{A dataframe to look for condition}

\item{cond}{A condition to check}

\item{text}{The message to print (if empty the condition is used)}

\item{threshold}{Cutoff number for ID reporting, as number}

\item{varname}{Column name of ID to print}

\item{na.rm}{Remove missing ID in the list to print, by default is False}

\item{data_old}{A dataset where to look for previously checked values}
}
\value{
Message to print as list
}
\description{
printIf2
prints the list of IDs meeting the specified condition
}
\examples{
df <- data.frame(ID = 1:4, Vaccs = c("pfizer"," ", "pfizer", "moderna"))
printIf(data = df, cond = Vaccs == "pfizer", threshold = 30, text = "Pfizer vaccin", varname = "ID")
}
