% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCheck.R
\name{printKableIf}
\alias{printKableIf}
\title{printKableIf
Takes any number of lists of caption-condition pairs, and vector of headers to match each list.
Outputs a table sectioned / list with header.}
\usage{
printKableIf(
  data,
  parameters,
  header = "Data Check",
  threshold = NULL,
  varname = "id",
  na.rm = FALSE,
  data_old = NULL
)
}
\arguments{
\item{data}{}

\item{parameters}{}

\item{header}{}

\item{threshold}{}

\item{varname}{}

\item{na.rm}{}

\item{data_old}{}
}
\value{
table
}
\description{
printKableIf
Takes any number of lists of caption-condition pairs, and vector of headers to match each list.
Outputs a table sectioned / list with header.
}
\examples{
parameter <- list(
"!is.na(SariWho) & SariWho==0 & ((!is.na(SympFever)&SympFever==0) | (!is.na(SympCough) & SympCough==0))" = "**Not meeting SARI case definition due to not having fever or cough symptoms**"
, "!is.na(SariWho)&SariWho==0&is.na(Dischargedate)&!is.na(Outcome)&Outcome==2" = "**Not meeting SARI case definition due to missing hospital discharge date when outcome is discharged**"
, "!is.na(SariWho)&SariWho==0&is.na(Deathdate)&!is.na(Outcome)&Outcome==1" = "**Not meeting SARI case definition due to missing hospital death date when outcome is deceased**"
, "!is.na(SariWho)&SariWho==0&is.na(Transferdate)&!is.na(Outcome)&Outcome==4" = "**Not meeting SARI case definition due to missing hospital transfer date when outcome is transferred.**"
)
}
