% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printVar}
\alias{printVar}
\title{printVar}
\usage{
printVar(dataset, pattern, regex = FALSE)
}
\arguments{
\item{dataset}{A data.frame containing the variable to be listed}

\item{pattern}{A pattern for varname in the dataframe see \code{regex}}

\item{regex}{Should the pattern be used as regex expression or use classical "joker" ? and * , the default}
}
\value{
A data frame of 10 rows with selected columns
}
\description{
Function used to quickly print/list a serie of variable (similar names like vaccdate, vacctype, vaccbrand )
}
\examples{
 data <- data.frame(Id = 1:4 ,
                   vaccage = c(34,45, 50,22 ),
                   symp = c("Y","Y","N","N"),
                   vaccboost=c("N","Y","N","Y"))
printVar(data,"symp")
printVar(data,"vac*")

}
