% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RefSite}
\alias{RefSite}
\title{Dataset correspondence table between country/study site codes, country/study site names}
\format{
A data frame with 40 rows and 5 variables:
\describe{
\item{ID}{Code associated to a location (country or study site), e.g. AT, BE, BG, etc.}
\item{Label}{Full name of the country or study site including
'the' article for NL and UK  e.g. Austria, Belgium, the Netherlands, the United Kingdom etc.}
\item{LocationName}{Shorter name of the country or study site}
\item{LocationType}{For location, variable specifying the goegraphical level, i.e., country or Nuts3 location}
\item{CountryISO2Code}{ISO2 code of the corresponding country, e.g. ES for Navarra}
}
}
\usage{
RefSite
}
\description{
Dataframe providing the correspondence table of the id code \code{ID}
and the geographical label \code{Label} to use
throughout reports. Additional information is available in columns
\code{LocationName},
\code{LocationType},
and \code{CountryISO2Code}
}
\keyword{datasets}
