% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename}
\alias{rename}
\title{Change the name of a data.frame column}
\usage{
rename(data, oldname, newname)
}
\arguments{
\item{data}{A data frame passed to the function}

\item{oldname}{Name of the column/variable to rename}

\item{newname}{New name to apply}
}
\value{
Nothing, the passed data.frame is modified directly
}
\description{
Simple function to rename a column in a data.frame
The rename is done "in place", no need to reassign the data.frame
A message is printed to confirm the change
}
\details{
Data.frame and column names are passed to the function as symbols (without "")
}
\examples{
df <- as.data.frame( c(One=1,Two=2) )
rename(df,Two,Last)
}
