% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{replaceInExcelAll}
\alias{replaceInExcelAll}
\title{Batch Find and Replace in Excel Files within a Directory}
\usage{
replaceInExcelAll(dir_path, oldname, newname, word = FALSE, ignore.case = TRUE)
}
\arguments{
\item{dir_path}{The directory path where Excel (.xlsx) files are located.}

\item{oldname}{The text to search for in the Excel files.}

\item{newname}{The text that will replace \code{oldname}.}

\item{word}{Logical, if TRUE only whole words will be replaced.}

\item{ignore.case}{Logical, if TRUE the function ignores case.}
}
\value{
No return value; the function prints a message indicating that all files have been modified.
}
\description{
This function performs a find and replace operation for specified text across all Excel (.xlsx) files in a given directory.
You can specify whether to replace only whole words and whether to perform case-sensitive or case-insensitive replacements.
}
\examples{
\dontrun{
replaceInExcelAll("./excelFiles/", "oldText", "newText")
}
}
