% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replicate}
\alias{replicate}
\title{replicate}
\usage{
replicate(char, ntime)
}
\arguments{
\item{char}{Character to replicate}

\item{ntime}{Number of replication}
}
\value{
\code{char} replicated \code{ntime}
}
\description{
replicate
}
\examples{
replicate("-",60)

}
