% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replichar}
\alias{replichar}
\title{replichar}
\usage{
replichar(char, ntime)
}
\arguments{
\item{char}{Character to replicate}

\item{ntime}{Number of replication}
}
\value{
\code{char} replicated \code{ntime}
}
\description{
replichar
}
\examples{
replichar("-",60)

}
