% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReverseFactorUsingDico.R
\name{reverseFactorUsingDico}
\alias{reverseFactorUsingDico}
\title{reverseFactorUsingDico}
\usage{
reverseFactorUsingDico(data, dictionary = NULL, dicos = NULL)
}
\arguments{
\item{data}{A dateset to process}

\item{dictionary}{A dictionary epiuf structure}

\item{dicos}{A dicos epiuf structure}
}
\value{
The dataset modified
}
\description{
reverseFactorUsingDico
}
