% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{right}
\alias{right}
\title{Extract x right characters of a string}
\usage{
right(text, num_char)
}
\arguments{
\item{text}{Text to extract from}

\item{num_char}{Number of char to extract from right}
}
\value{
\code{num_char} extracted characters from right side
}
\description{
This function allow to extract the x last character of a string
(Character are counted from the end)
It could be used to extract the last part of a string
To retrieve extension of a file use \code{fileExt} instead
}
\examples{
right("dummy_example",7)
right("data_completed_2022",4)


}
