% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{saveDictionary}
\alias{saveDictionary}
\title{saveDictionary}
\usage{
saveDictionary(filename = NULL, dictionary = NULL)
}
\arguments{
\item{filename}{The file name and path where to save the dictionary (character).
Any existing file will be overwritten.}

\item{dictionary}{(Optional) Dictionary. By default the current dictionary will be saved.}
}
\value{
No return value
}
\description{
Function to save the current dictionary \code{'dictionary'} as an Excel
spreadsheet in a given folder \code{'filename'}.
}
\examples{
\dontrun{
saveDictionary(filename = paste0(getwd(), "/test.xlsx"))
}

}
\seealso{
\code{\link[=getDictionary]{getDictionary()}}
}
