% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasourcestools.R
\name{saveSourceList}
\alias{saveSourceList}
\title{saveSourceList}
\usage{
saveSourceList(filename = NULL)
}
\arguments{
\item{filename}{Name of the file containing sources informations}
}
\value{
nothing
}
\description{
saveSourceList
}
\examples{
\dontrun{
saveSourceList("datasourcelist.xls")
}
}
