% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{saveXlsx}
\alias{saveXlsx}
\title{saveXlsx}
\usage{
saveXlsx(filename = "", wb = NULL)
}
\arguments{
\item{filename}{Specify an optional filename (xlsx). If empty the current name of the workbook will be used.}

\item{wb}{An optional workbook created or opened by xlsx.
If not specified, the last workbook loaded with openXlsx will be used.}
}
\value{
Saved file
}
\description{
This function saves the current workbook as a filename.
}
\examples{

# Creating an empty workbook
wb <- openxlsx::createWorkbook()
sheetname <- "First Sheet"
openxlsx::addWorksheet(wb, sheetName = sheetname)

# Writing a data frame in the workbook 
fillCells(wb = wb, 
          onesheet = sheetname, 
          line = 1, 
          col = 1, 
          iris)
          
# Saving the Excel file
\dontrun{
saveXlsx(filename = "test.xlsx", wb = wb)
}

}
\seealso{
\code{\link[=openXlsx]{openXlsx()}}
}
