% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasourcestools.R
\name{searchNewFiles}
\alias{searchNewFiles}
\title{searchNewFile}
\usage{
searchNewFiles(path, country = "", pattern = NULL)
}
\arguments{
\item{path}{Path to serach for (relative to the path where is the datasources file)}

\item{country}{Country code to attach to the file information}

\item{pattern}{Optional pattern to restrict file list}
}
\value{
List of new files (file names)
}
\description{
Part of datasources management
Look into a folder (usually a country folder) to find new or updated datasources
If such datasources are found, they are imported in the datasourcelist with NEW or UPDATED status and current date
DatasourceList contain one line for each new event
}
\examples{
\dontrun{ searchNewFiles("sourcesdir","FR") }
}
