% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maindialog.R
\name{setDialog}
\alias{setDialog}
\title{setDialog}
\usage{
setDialog(dialog, value = NULL)
}
\arguments{
\item{dialog}{A name corresponding to a variable in the dialog Env}

\item{value}{The value to assign to that variable}
}
\value{
Nothing
}
\description{
setDialog
}
