% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{setDictionary}
\alias{setDictionary}
\title{setDictionary}
\usage{
setDictionary(dictionary)
}
\arguments{
\item{dictionary}{A dictionary epiuf structure (data.frame)}
}
\value{
No return value
}
\description{
Set a dictionary using a data.frame (this function would be rarely used except for tests).
}
\examples{

# Create an example data frame to set as a dictionary 
dic <- data.frame(generic_name = c("id", "idcountry", "age", "agegroup3", "agroup10"),
                  source_name = c("id", "idcountry", "age", "agegroup3", "agroup10"),
                  type = c("character", "character", "numeric", "factor", "factor"),
                  dico = c("","","","agegp3","agegp10"))

# Set the dictionary 
setDictionary(dictionary = dic)


}
\seealso{
\code{\link[=getDictionary]{getDictionary()}}, \code{\link[=openDictionary]{openDictionary()}}, \code{\link[=applyDictionary]{applyDictionary()}}
}
