% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{setDictionaryActions}
\alias{setDictionaryActions}
\title{setDictionaryActions}
\usage{
setDictionaryActions(actions)
}
\arguments{
\item{actions}{A dataset (data frame) of actions}
}
\value{
No return value
}
\description{
Set the data frame \code{actions} as a dictionary action in the environment
as returned by \code{getNewDictionaryLine("actions")}
(i.e., columns: 'variable', 'action_group', 'parameters').
}
\examples{
# Example of empty dictionary action
getNewDictionaryLine(mode = "actions")

# Set it as dictionary action in environment
setDictionaryActions(getNewDictionaryLine(mode = "actions"))


 
}
\seealso{
\code{\link[=getDictionaryActions]{getDictionaryActions()}}, \code{\link[=getNewDictionaryLine]{getNewDictionaryLine()}}
}
