% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiOption.R
\name{setEpiOption}
\alias{setEpiOption}
\title{setEpiOption}
\usage{
setEpiOption(EpiOption, value)
}
\arguments{
\item{EpiOption}{name of the option to assign as character}

\item{value}{The value to be assigned to option}
}
\value{
The previous option value before new assignment
}
\description{
assign a package option
}
\examples{
setEpiOption("option",1)


}
