% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOrGetFilePath.R
\name{setOrGetFilePath}
\alias{setOrGetFilePath}
\title{Set or get a file path for REDCap data downloads}
\usage{
setOrGetFilePath(country)
}
\arguments{
\item{country}{A country code for a REDCap country project}
}
\value{
The file path within which downloaded REDCap data will be placed
}
\description{
This function walks the user through setting the filepath that can be called automatically when downloading REDCap data
}
\examples{
\dontrun{
accessREDCap("DB")
# If file path is not found for this country, setOrGetFilePath will then run
}
}
