% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasourcestools.R
\name{setSourceList}
\alias{setSourceList}
\title{setSourceList}
\usage{
setSourceList(datasourcedf)
}
\arguments{
\item{datasourcedf}{a dataframe containing sources information}
}
\value{
nothing
}
\description{
setSourceList
}
\examples{
\dontrun{ setSourceList(dfsourcelist) }
}
