% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortby.R
\name{sortBy}
\alias{sortBy}
\title{sortBy
Sort a dataframe according to a field optionnaly decreasing}
\usage{
sortBy(data, field, decreasing = FALSE, na.last = TRUE)
}
\arguments{
\item{data}{The dataframe to besorted}

\item{field}{The field on which to sort (numeric or character)}

\item{decreasing}{Default is FALSE}
}
\value{
The sorted dataframe
}
\description{
sortBy
Sort a dataframe according to a field optionnaly decreasing
}
\examples{
df <- data.frame(one=c(3,1,5,2))
df <- sort(df,"one")
df <- sort(df,1) 
}
