% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{sourceFile}
\alias{sourceFile}
\title{sourceFile}
\usage{
sourceFile(pathname, filename)
}
\arguments{
\item{pathname}{The label given to a previously specified path. Needs to be written in "".}

\item{filename}{The name of the file to source (if it contains a path, it will be added to the specified path (pathname).}
}
\value{
nothing
}
\description{
'sourceFile()' sources a file from a previously set path and runs it in your script.
It uses the label of the path (pathname) and the name of the file (filename) of which you want to run.
This function is a wrapper for 'source()'.
}
\examples{

## Not to run - example only 
\dontrun{
# Set an example path named 'sources' 
setPath(pathname = "SOURCES", 
       path = "c:/dev/Resources", 
       makedir = "Never")

# Create an example file 
file <- tempfile(fileext = ".R")

# Source example file from the specified path 
sourceFile("SOURCES",file)
}
}
\seealso{
\code{\link[=setPath]{setPath()}}
}
