% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasourcestools.R
\name{sourceSummary}
\alias{sourceSummary}
\title{sourceSummary}
\usage{
sourceSummary(country, startdate = NULL, enddate = NULL)
}
\arguments{
\item{country}{For specific conutry table}

\item{startdate}{First date}

\item{enddate}{Last date}
}
\value{
a dataframe ready to print
}
\description{
Produce a table of status by country/period
}
\examples{
\dontrun{ sourceSummary("FR")}
}
