% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyring.r
\name{storeKeyring}
\alias{storeKeyring}
\title{Store a secret in a new keyring}
\usage{
storeKeyring(keyring_id, secret)
}
\arguments{
\item{keyring_id}{A string specifying the keyring ID.}

\item{secret}{A string representing the secret to be stored in the keyring.}
}
\value{
A string with the stored secret.
}
\description{
This function creates a new keyring and stores a provided secret in it.
}
