% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{tab}
\usage{
tab(..., includeNA = "ifany", data)
}
\arguments{
\item{...}{Variable or vector.}

\item{includeNA}{Specify whether NAs should be included in the results or not.}

\item{data}{The dataset which contain the variables (data.frame).}
}
\value{
cross table
}
\description{
Function to easily output tabs and cross tabs of variables - wrapper for
the table() function. Requires dply.
}
