% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{txtFindReplace}
\alias{txtFindReplace}
\title{Find and Replace Text in a Text File}
\usage{
txtFindReplace(
  filename,
  pattern,
  replacement,
  wholeword = TRUE,
  ignore.case = FALSE,
  listonly = FALSE
)
}
\arguments{
\item{filename}{Character. The name of the file to modify.}

\item{pattern}{Character vector. A vector of patterns to search for in the file.}

\item{replacement}{Character vector. A vector of replacements for the patterns found.}

\item{wholeword}{Logical. If TRUE, only whole words will be replaced. Defaults to TRUE.}

\item{ignore.case}{Logical. If TRUE, the function performs a case-insensitive search. Defaults to FALSE.}

\item{listonly}{Logical. If TRUE, the function only lists the files that would be modified without actually
modifying them. Defaults to FALSE.}
}
\value{
NULL. The function modifies the text file in-place and prints the exact number of changes made.
}
\description{
This function reads a text file and replaces specified patterns with given replacements.
It allows for multiple patterns and replacements, and can perform case-sensitive or
case-insensitive searches. It can also match whole words.
}
\examples{
\dontrun{
txtFindReplace("example.txt", "oldText", "newText")
txtFindReplace("example.txt", c("oldText1", "oldText2"), "newText")
}
}
\seealso{
\code{\link{fileFindReplace}}, \code{\link{xlsxFindReplace}}
}
