% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDataSet.R
\name{updateDataset}
\alias{updateDataset}
\title{UpdateDataset}
\usage{
updateDataset(data, modele)
}
\arguments{
\item{data}{A dataset to update}

\item{modele}{A dataset used as model for the final structure}
}
\value{
the updated dataset with modele structure
}
\description{
Transform a dataset by  adding/removing column according to a model dataset
}
