% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maindialog.R
\name{updatefilelist}
\alias{updatefilelist}
\title{updatefilelist  Used internally by maindialog to refresh file list}
\usage{
updatefilelist()
}
\value{
The list of file extracted from datasource for the given criteria
criteria are read from dialog values
}
\description{
updatefilelist  Used internally by maindialog to refresh file list
}
