% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateOrDeleteKeyring.R
\name{UpdateOrDeleteKeyring}
\alias{UpdateOrDeleteKeyring}
\title{Delete or Update a Keyring}
\usage{
UpdateOrDeleteKeyring(keyring_id)
}
\arguments{
\item{keyring_id}{A character string specifying the keyring name to delete or update}
}
\value{
A character string containing the password for further API calls.
}
\description{
This function checks if the specified keyring exists, and if so, asks the user whether they'd like to delete or update it.
}
\examples{
\dontrun{
# Set a secret for a fictional country called 'DB'
UpdateOrDeleteKeyring("DB")
}
}
