% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyspelling.R
\name{verifySpelling}
\alias{verifySpelling}
\title{verifySpelling}
\usage{
verifySpelling(varname, CorrectList, ErrPerc = 0.25, verbose = TRUE)
}
\arguments{
\item{varname}{A string or a string vector which will be compared to another string
or a list of strings in order to find similarity or sound like.
The objective is to automatically correct small mistake in wording.}

\item{CorrectList}{A string or a list of string from which \code{varname} will be searched.}

\item{ErrPerc}{An acceptable errPerc when comparing string , default to 0.10\%.}

\item{verbose}{If TRUE, a message is issued.}
}
\value{
The string guessed from varname using CorrectList.
}
\description{
This function is used for testing a word against a list. It then returns the most similar or NULL if there is nothing.
}
\examples{

# Verify the spelling of January
verifySpelling("Janury",c("January", "Janvier" ))

}
