% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewIf.R
\name{viewIf}
\alias{viewIf}
\title{List head of data.frame allowing to select column to view}
\usage{
viewIf(data, ..., cond = NULL, nline = 10)
}
\arguments{
\item{data}{The dataset to explore}

\item{...}{variable name(s) to include in view}

\item{cond}{A logical condition to select records}

\item{nline}{Number of rows to show (default to n) If negative then number of row from the end}
}
\value{
invisibly, the resulting selection
}
\description{
List head of data.frame allowing to select column to view
}
\examples{
\dontrun{
view(data ,age,case)
view(data, ,case, nline=5)
view(data, age,case, nline= -10)
}
}
