% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoSendsThis.R
\name{whoSendsThis}
\alias{whoSendsThis}
\title{whoSendsThis}
\usage{
whoSendsThis(
  searchterm,
  filetags,
  dictionary_location,
  dictionary_root_filename
)
}
\arguments{
\item{searchterm}{The character string which wish to look}

\item{filetags}{The list of dictionary filename endings to search through}

\item{dictionary_location}{file path to where all the dictionaries that you want to search through are stored}

\item{dictionary_root_filename}{root filename for all the dictionaries that wish to search through, which change only by the filetag at the end}
}
\value{
data.frame
}
\description{
Description
The aim of this function to is to quickly ascertain which dictionaries- in a
specified folder contain your searchterm.
}
\details{
This function looks into the specified file location, and then opens each dictionary
in turn as specified by the dictionary_root_filename and filetags. Once opened it
performs a grep search for the searchterm in the source_name and the generic_name
colunms and lists all matching variablenames in a table.
}
\examples{
\dontrun{ whoSendsThis("covaccin",c("country","other"),"c:/root_dico_dir","diconameroot") }
}
\seealso{
For more details see the link below to access the vignette:
\href{../doc/epiuf_package.html}{\code{vignette("epiuf_package")}}
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
