% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{xlsxFindReplace}
\alias{xlsxFindReplace}
\title{Find and Replace Text in an Excel File}
\usage{
xlsxFindReplace(xlsxName, oldname, newname, word = FALSE, ignore.case = TRUE)
}
\arguments{
\item{xlsxName}{The name (including path) of the Excel (.xlsx) file to modify.}

\item{oldname}{The text to search for in the Excel file.}

\item{newname}{The text that will replace \code{oldname}.}

\item{word}{Logical, if TRUE only whole words will be replaced.}

\item{ignore.case}{Logical, if TRUE the function ignores case.}
}
\value{
No return value; the function modifies the Excel file in-place and prints a message indicating the modified file.
}
\description{
This function reads an Excel file and replaces specified text (\code{oldname}) with a new text (\code{newname}) across all sheets.
You can specify whether to replace only whole words and whether to perform case-sensitive or case-insensitive replacements.
}
\examples{
\dontrun{
xlsxFindReplace("example.xlsx", "oldText", "newText")
}
}
