% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yesno}
\alias{yesno}
\title{yesno}
\usage{
yesno(message)
}
\arguments{
\item{message}{The prompt to display before the Yes/No choice}
}
\value{
True if yes, FALSE if no and NA otherwise
}
\description{
Dispkay a prompt and wait for a Yes/No answer
}
