% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipGithubRepo.R
\name{zipGithubRepo}
\alias{zipGithubRepo}
\title{zipGithubRepo}
\usage{
zipGithubRepo(
  zipfile_destination,
  zipfile_name,
  commit = TRUE,
  commit_msg = NULL,
  pull = TRUE,
  push = TRUE
)
}
\arguments{
\item{zipfile_destination}{Where you want to save your zipped repository}

\item{zipfile_name}{What you want to call your saved zipped repository}

\item{commit}{Logical - do you want to make a commit of your repository before extraction?}

\item{commit_msg}{If commit, what is your message. Default is: commit pre zip (todays date)}

\item{pull}{Logical - do you want to make a pull from github before extraction?}

\item{push}{Logical - do you want to make a push to github before extraction?}
}
\description{
Description
Make an extraction from your current github repo and saves it to the indicated
folder under the indicated name. Options to make a comit and pull from then
push to github before extracting the repository.
}
\seealso{
For more details see the link below to access the vignette:
\href{../doc/epiuf_package.html}{\code{vignette("epiuf_package")}}
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
