% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_month_year_tools.R
\name{abrvMonth}
\alias{abrvMonth}
\title{abrvMonth}
\usage{
abrvMonth(date)
}
\arguments{
\item{date}{A date Value}
}
\value{
The month of the date in abreviated format
}
\description{
abrvMonth
}
\examples{
abvMonth(Sys.Date())
}
