% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{abrvMonth}
\alias{abrvMonth}
\title{abrvMonth}
\usage{
abrvMonth(date)
}
\arguments{
\item{date}{A date Value}
}
\value{
The month of the date in abbreviated format
}
\description{
Returns the month in abbreviated format of a date object
}
\examples{
abrvMonth(Sys.Date())

}
\seealso{
\code{\link[=Month]{Month()}} and \code{\link[base:format]{base::format()}}
}
