% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addSep}
\alias{addSep}
\title{addSep
transform a list into a string of separated values}
\usage{
addSep(li, c)
}
\arguments{
\item{li}{A list of values}

\item{c}{A separator}
}
\value{
A string representing the list as separated values
}
\description{
addSep
transform a list into a string of separated values
}
\examples{
li <- list("one","two","three")
addSep(li,",")
}
