% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{applyDictionary}
\alias{applyDictionary}
\title{applyDictionary}
\usage{
applyDictionary(dictionary = NULL, data, verbose = TRUE, keepextra = FALSE)
}
\arguments{
\item{dictionary}{A dictionary (data.frame)}

\item{data}{A dataset to transform to generic structure of the dictionary (data.frame)}

\item{verbose}{Feedback regarding matching of the two}

\item{keepextra}{Logical. If TRUE, extra variables existing in data
but not in generic dictionary are kept in the returned dataset.
Then this generic dataset is no longer generic because it may contain
non generic variables.}
}
\value{
A data set
}
\description{
Updates dataset \code{data} and transform it according to the dictionary generic structure.
}
\examples{
 
# Create dummy dictionary
dic <- data.frame(generic_name = c("id", "idcountry", "age", "agegroup3", "agroup10"),
                  source_name = c("ID", "CountryID", "Age", "Agegp3", "Agep10"),
                  type = c("character", "character", "numeric", "factor", "factor"),
                  dico = c("","","","agegp3","agegp10"),
                  unknowns = NA,
                  description = NA,
                  comments = NA)

# Set the current dictionary 
setDictionary(dictionary = dic)

# Create corresponding dummy dicos  
dicos <- data.frame(dico_name = c("yesno", "yesno", "agegp3", "agegp3", "agegp3"),
                    label = c("no", "yes", "0-14 years", "15-64 years", "65+ years" ),
                    code = c("0", "1", "0", "1", "2"))

# Set the corresponding dicos
setDicos(dicos)

# Dummy source data
source <- data.frame(ID = c(1,2,3),
                     CountryID = c(2,3,4),
                     Age = c(4,5,6))

# Transforming the source data into generic
new <- applyDictionary(dictionary = dic,
                       data = source)
                       
# Example base on epiuf::DummyData          
new <- applyDictionary(dictionary = dic, 
                       data = epiuf::DummyData)
new <- applyDictionary(dictionary = dic, 
                       data = epiuf::DummyData,
                       keepextra = TRUE)
               

}
