% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{applyDictionary}
\alias{applyDictionary}
\title{applyDictionary}
\usage{
applyDictionary(dictionary = NULL, data, verbose = TRUE, keepextra = FALSE)
}
\arguments{
\item{dictionary}{A dictionary (passed as dataframe)}

\item{data}{A dataset to transform to generic}

\item{verbose}{Should we have feedback}

\item{keepextra}{if TRUE, extra variables existing in data but not in generic dictionary are keep
in the returned dataset. Then this generic dataset is no longer generic because it may contain
non generic variables}
}
\value{
A data set
}
\description{
applyDictionary
}
