% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyNa.R
\name{applyNA}
\alias{applyNA}
\title{Title}
\usage{
applyNA(data, varname, searchlist = NULL, join = FALSE)
}
\arguments{
\item{data}{}

\item{varname}{}

\item{searchlist}{}

\item{join}{}
}
\value{

}
\description{
Title
}
