% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyNa.R
\name{applyNA}
\alias{applyNA}
\title{applyNA}
\usage{
applyNA(data, varname, searchlist = NULL, join = FALSE)
}
\arguments{
\item{data}{Dataset which contain the variable to be processed}

\item{varname}{The variable where all kind of UNKNOWNS should be identified and replaced by NA}

\item{searchlist}{An optional list of unknown values}

\item{join}{False by default, if TRUE, the searchlist will be added to the default list of UNKNOWNS}
}
\value{
variable with NAs edited and message to confirm, or warning   GDE or data ?
}
\description{
Default Unknowns : dnk nd na ne zna unknown do not know
}
