% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyNAAll.R
\name{applyNAAll}
\alias{applyNAAll}
\title{applyNAAll}
\usage{
applyNAAll(data, dictionary = NULL, join = TRUE)
}
\arguments{
\item{data}{The dataset where to applyNA for each variable using the dictionnary}

\item{dictionary}{an optional dictionary (epiuf strucure see \link{getDictionary})}

\item{join}{Should default NA value added to those from dictionary}
}
\value{
data.frame and printed output
}
\description{
applyNAAll
}
