% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyNAAll.R
\name{applyNAAll}
\alias{applyNAAll}
\title{applyNA}
\usage{
applyNAAll(data, dictionary = NULL)
}
\arguments{
\item{data}{}

\item{dictionary}{(optional)}
}
\value{
data.frame and printed output
}
\description{
applyNA
}
\examples{
df <- applyNAAll(df)


}
