% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{calYear}
\alias{calYear}
\title{calYear}
\usage{
calYear(date)
}
\arguments{
\item{date}{A date value}
}
\value{
The calendar Year of the date as numeric
}
\description{
Extract the year of a date object and return it as numeric
}
\examples{
calYear(Sys.Date())

}
\seealso{
\code{\link[=isoYear]{isoYear()}} and \code{\link[base:format]{base::format()}}
}
