% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_month_year_tools.R
\name{calYear}
\alias{calYear}
\title{calYear}
\usage{
calYear(date)
}
\arguments{
\item{date}{A date value}
}
\value{
The calendar Year of the date
}
\description{
calYear
}
\examples{
calYear(Sys.Date())
}
