% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{camel}
\alias{camel}
\title{camel
To transform snake_case into CamelCase}
\usage{
camel(x)
}
\arguments{
\item{x}{a string or a sring list or a vector of strings}
}
\value{
a CamelCase value or list
}
\description{
camel
To transform snake_case into CamelCase
}
\examples{
test <- "snake_case"
camel(test)

}
