% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catret}
\alias{catret}
\title{Catret}
\usage{
catret(...)
}
\arguments{
\item{...}{A list of values to be concatenated for console output. Direct text should be written in "" and seperated using ','. See examples.}
}
\value{
Text in the console.
}
\description{
Catret concatenates the provided entry and prints the result in the console. Any text function (e.g. catret or cat) used after catret will be printed on a new line.
}
\details{
Catret is a wrapper for the base function 'cat'. It takes any number of arguments.
}
\examples{
# Simple text 
catret("A line of simple text")

# Printing text and a value found in the environment
site <- 'ES'
catret("The site", site, "has a large population")

# Example of catret printing the next function on a new line 
{catret("elephant"); catret("cow"); cat("dog")}

}
\seealso{
cat()
}
