% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catret}
\alias{catret}
\title{catret
cat ret is a wrapper for cat(...,"newline")}
\usage{
catret(...)
}
\arguments{
\item{...}{list of values to concatenate for console output}
}
\value{
nothing
}
\description{
catret
cat ret is a wrapper for cat(...,"newline")
}
\examples{
catret("test")

}
