% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catret}
\alias{catret}
\title{catret}
\usage{
catret(...)
}
\arguments{
\item{...}{list of values to concatenate for console output}
}
\value{
nothing
}
\description{
catret is a wrapper for cat(...,"newline").

catret concatenate all provided entry and
output the result to the console, then out a carriage return to make any further cat
function to start on the next line.
}
\examples{
catret("test")

}
