% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charCount}
\alias{charCount}
\title{count how many char}
\usage{
charCount(pattern, stosearch)
}
\arguments{
\item{pattern}{"The character or pattern to search for}

\item{stosearch}{The string to search in}
}
\value{
number of match
}
\description{
Count the number of a specified char into a text using reg expr
If pattern is a single character or a suite of character, then charcount return
the number of occurrence of pattern into "tosearch"
}
\details{
If pattern is a regular expression, charCount return the number of time this regular
expression is verified into "tosearch" string
Verify regular expression syntax for specific exptression like . which should
be escaped by \\ : \\.
}
\examples{
nb <- charCount("/", "test/essai/try")
nb <- charCount("^t", "test/essai/try")
nb <- charCount(".t", "test/essai/try")


}
\seealso{
\code{\link[=nchar]{nchar()}}
}
