% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{charCount}
\alias{charCount}
\title{Count the Occurrences of a Pattern in Text}
\usage{
charCount(pattern, stosearch)
}
\arguments{
\item{pattern}{Character. The pattern to search for within the text string.}

\item{stosearch}{Character. The text string or vector in which to search for the pattern.}
}
\value{
Integer. The total count of occurrences of the pattern in the text string.
}
\description{
This function counts the number of times a specified pattern appears in a given text string.
The function uses regular expression matching and returns the total count of occurrences.
If pattern is a single character or a suite of character, then charcount return
the number of occurrence of pattern into "tosearch"
}
\details{
If pattern is a regular expression, charCount return the number of time this regular
expression is verified into "tosearch" string
Verify regular expression syntax for specific expression like . which should
be escaped by \\ : \\.
}
\examples{

charCount("a", "banana")  # Returns 3 (three 'a's in "banana")
charCount("a", c("banana","apple"))  # Returns 4 
charCount("\\\\d", "13ab2c")  # Returns 3 (three digits in "13ab2c")

}
\seealso{
\code{\link[=nchar]{nchar()}}
}
