% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDicoAll.R
\name{checkDicoAll}
\alias{checkDicoAll}
\title{Title}
\usage{
checkDicoAll(data, dictionary = NULL, dicos = NULL)
}
\arguments{
\item{data}{}

\item{dictionary}{}

\item{dicos}{}
}
\value{

}
\description{
Title
}
