% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDicoAll.R
\name{checkDicoAll}
\alias{checkDicoAll}
\title{checkDicoAll}
\usage{
checkDicoAll(data, dictionary = NULL, dicos = NULL)
}
\arguments{
\item{data}{Dataset to process}

\item{dictionary}{An optional dictionary (epiuf strucure see \link{getDictionary})}

\item{dicos}{A Dico structure containing code and labels}
}
\value{
text outputs
}
\description{
checkDicoAll
}
