% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkType.R
\name{checkType}
\alias{checkType}
\title{checkType
Gives feedback on whether a variable matches the type specified.}
\usage{
checkType(data, varname, type = c("character", "date", "factor", "numeric"))
}
\arguments{
\item{data}{The dataset}

\item{varname}{The var}

\item{type}{The expeted type}
}
\value{
report
}
\description{
checkType
Gives feedback on whether a variable matches the type specified.
}
\examples{
\dontrun{
   checkType(data,var)
}

}
