% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTypeAll.R
\name{checkTypeAll}
\alias{checkTypeAll}
\title{checkTypeAll}
\usage{
checkTypeAll(data, dictionary = NULL, flag = NULL)
}
\arguments{
\item{data}{The dataset (data.frame).}

\item{dictionary}{The epiuf dictionary.}

\item{flag}{A specific class of interest.}
}
\value{
report
}
\description{
Uses the dictionary to compare the class of each variable to the
type listed. A specific class of interest can be specified with the flag
parameter.
}
