% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTypeAll.R
\name{checkTypeAll}
\alias{checkTypeAll}
\title{checkTypeAll
Uses dictionary to compare class of each variable to type listed.
A specific class of interest can be specified with the flag parameter}
\usage{
checkTypeAll(data, dictionary = NULL, flag = NULL)
}
\arguments{
\item{data}{}

\item{dictionary}{}

\item{flag}{}
}
\value{
report
}
\description{
checkTypeAll
Uses dictionary to compare class of each variable to type listed.
A specific class of interest can be specified with the flag parameter
}
