% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanNames}
\alias{cleanNames}
\title{cleanNames}
\usage{
cleanNames(name, chr = "")
}
\arguments{
\item{name}{A string to clean by removing non ascii char and accents}

\item{chr}{A chr used to replace non ascii character}
}
\value{
the cleaned string
}
\description{
cleanNames
}
\examples{
x <- "ÁbcdêãçoàúüEssai/=+$67"
cleanNames(x)  # no libraries needed
}
