% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseVar.R
\name{collapseVar}
\alias{collapseVar}
\title{collapsVar
Function aims to take checkbox variables that are not expanded, and retain only
one input / patient based off hirearchy of input. Hierarchy could be derrived from dico codes
set in specific order. Idea is that all checkboxes can be given action group
tags depending on if they are to be expanded or collapse.
collapseVar(data, varname, code hirearchy) (take hirearchy from dicolist - put in desired order manually) - is that dangerous????
\itemize{
\item get separated string list for each input.
\item if length is 1 - leave
\item if length is >1 - assess
\item retain highest value from the order of preference
}}
\usage{
collapseVar(data, mydictionary, multivarname)
}
\arguments{
\item{data}{The dataset which contain a multivar}

\item{mydictionary}{dico to use (or ordered list ?)}

\item{multivarname}{Name of multi variable}
}
\value{
The multi variable collapsed
}
\description{
collapsVar
Function aims to take checkbox variables that are not expanded, and retain only
one input / patient based off hirearchy of input. Hierarchy could be derrived from dico codes
set in specific order. Idea is that all checkboxes can be given action group
tags depending on if they are to be expanded or collapse.
collapseVar(data, varname, code hirearchy) (take hirearchy from dicolist - put in desired order manually) - is that dangerous????
\itemize{
\item get separated string list for each input.
\item if length is 1 - leave
\item if length is >1 - assess
\item retain highest value from the order of preference
}
}
