% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseVarAll.R
\name{collapseVarAll}
\alias{collapseVarAll}
\title{collapseVarAll}
\usage{
collapseVarAll(data, action = NULL)
}
\arguments{
\item{data}{Data set to be processed (data.frame).}

\item{action}{A Dictionary action (list structure).}
}
\value{
The data set with collapsed variable.
}
\description{
This function applies a dictionary action to a data set.
}
\details{
It aims to take check box variables that are not expanded, and retain only
one input / patient based off hierarchy of input. Hierarchy could be derived from dico codes
set in specific order.

Idea is that all check boxes can be given action group tags depending on if they are to be expanded or collapse.
}
