% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseVarAll.R
\name{collapseVarAll}
\alias{collapseVarAll}
\title{collapseVarAll}
\usage{
collapseVarAll(data, action = NULL)
}
\arguments{
\item{data}{Dataset to be processed}

\item{action}{A Dictionary action list structure}
}
\value{
The dataset with collapsed variable
}
\description{
collapseVarAll
}
