% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printIf.R
\name{countIf}
\alias{countIf}
\title{Simple count of records satisfying  a conditional expression}
\usage{
countIf(data, cond = NULL)
}
\arguments{
\item{data}{The dataset where records should be counted}

\item{cond}{A logical expression}
}
\value{
A number of records
}
\description{
Simple count of records satisfying  a conditional expression
}
\examples{

df <- data.frame(Id = 1:4 ,
                    Vaccs = c("pfizer"," ", "pfizer", "moderna"))
nb <- countIf(df,Vaccs=='pfizer')
nb <- countIf(df)
}
