% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_month_year_tools.R
\name{countIsoWeeks}
\alias{countIsoWeeks}
\title{countIsoWeeks}
\usage{
countIsoWeeks(date = date, origin = "2020-10-05")
}
\arguments{
\item{date}{A date Value}

\item{origin}{A reference date value}
}
\value{
the nimber of isoweek between the two dates
}
\description{
countIsoWeeks
}
\examples{
countIsoWeeks(date = Sys.Date(),origin = "2000-01-01")
}
