% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{countIsoWeeks}
\alias{countIsoWeeks}
\title{countIsoWeeks}
\usage{
countIsoWeeks(date, origin = "2020-10-05")
}
\arguments{
\item{date}{A date Value}

\item{origin}{A reference date value in character string format}
}
\value{
the number of weeks between the two dates
}
\description{
Return the number of weeks between the two dates
}
\examples{
countIsoWeeks(date = Sys.Date(), origin = "2000-01-01")

}
\seealso{
\code{\link[base:difftime]{base::difftime()}}
}
\author{
Jenny Howard \email{j.howard@epiconcept.fr}
}
