% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{createXlsxStyle}
\alias{createXlsxStyle}
\title{createXlsxStyle}
\usage{
createXlsxStyle(...)
}
\arguments{
\item{...}{Parameters for createStyle}
}
\value{
a Style to be used in XLSX
}
\description{
createXlsxStyle
}
