% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{crosstab - outputs a cross table in format data frame}
\usage{
crosstab(data, var1, var2, missing = "always", extra = "None", decimals = 1)
}
\arguments{
\item{data}{The data set to look at}

\item{var1}{The first variable (the rows)}

\item{var2}{The second variable (the columns)}

\item{missing}{takes auguments 'no' to not show missing and 'always' to show missing (as for table())}

\item{extra}{What extra info you want. Current options "None", "Total" and "Percent". Default is "None". Feel free to add your own!}

\item{decimals}{state how many decimal points you want to list for the output of any calculations. Default is 1}
}
\description{
crosstab - outputs a cross table in format data frame
}
