% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{crosstab - outputs a cross table in format data frame}
\usage{
crosstab(data, var1, var2, missing = "always", extra = "None", decimals = 1)
}
\arguments{
\item{data}{The data set to look at}

\item{var1}{Character string, colomn name of the first variable (the rows)}

\item{var2}{Character string, colomn name of the second variable (the columns)}

\item{missing}{takes auguments 'no' to not show missing and 'always' to show missing (as for table())}

\item{extra}{What extra info you want. Current options "None", "Total" and "Percent". Default is "None". Feel free to add your own!}

\item{decimals}{state how many decimal points you want to list for the output of any calculations. Default is 1}
}
\description{
crosstab - outputs a cross table in format data frame
}
\examples{
data <- data.frame(id = 1:10,
                   cases = c(rep(1,3), rep(0,7)),
                   vacc = sample(c(0,1), replace = TRUE, size = 10))
table(data$cases, data$vacc, useNA = "always")
crosstab(data, 
         var1 = "cases", 
         var2 = "vacc")
crosstab(data, 
         var1 = "cases", 
         var2 = "vacc", 
         missing = "no")
crosstab(data, 
         var1 = "cases", 
         var2 = "vacc", 
         extra = "Total")
crosstab(data, 
         var1 = "cases", 
         var2 = "vacc", 
         extra = "Percent")

}
