% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{crosstab - outputs a cross table in format data frame}
\usage{
crosstab(data, var1, var2, missing = TRUE, extra = "Total")
}
\arguments{
\item{data}{The data set to look at}

\item{var1}{The first variable (the rows)}

\item{var2}{The second variable (the columns)}

\item{missing}{takes auguments 'no' to not show missing and 'always' to show missing (as for table())}

\item{extra}{What extra info you want. Currently default is 'Total' and output has row and column totals listed. There is space in the function for more kinds of extras to be added}
}

\description{
crosstab - outputs a cross table in format data frame
}
