% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCheck.R
\name{dataCheck}
\alias{dataCheck}
\title{dataCheck}
\usage{
dataCheck(
  data,
  data_old = NULL,
  cond,
  text = "",
  threshold = NULL,
  varname = "id"
)
}
\arguments{
\item{data}{}

\item{data_old}{}

\item{cond}{}

\item{text}{}

\item{threshold}{}

\item{varname}{}
}
\value{
a vector
}
\description{
dataCheck
}
