% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descBy.R
\name{descBy}
\alias{descBy}
\title{descBy}
\usage{
descBy(data, vars, labels = NULL, by = NULL, digits = 1)
}
\arguments{
\item{data}{name of the dataset}

\item{vars}{the variables list to describe (numeric or factor)}

\item{labels}{labels for the variables list}

\item{by}{variable to compare (must be a factor and have at least two categories)}

\item{digits}{number of digits to display (rounding)}
}
\value{
a descriptive table
}
\description{
Describe numeric variables or categorical variables (must be in factor) according or not to another variable in a table.
}
\examples{
 descBy(iris, "Sepal.Length", by = "Species", digits = 1) 


}
