% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descBy.R
\name{descBy}
\alias{descBy}
\title{descBy}
\usage{
descBy(data, vars, labels = NULL, compVar = NULL)
}
\arguments{
\item{data}{name of the dataset}

\item{vars}{the variables list to describe (numeric or factor)}

\item{labels}{labels for the variables list}

\item{compVar}{variable to compare (must be a factor and have at least two categories)}
}
\value{
a descriptive table
}
\description{
Aim: to describe numeric variables or categorical variables (must be in factor) according or not to another variable in a table.
}
\examples{
\dontrun{ 
   table <- descBy(data, vars = c("age","sex","fluvaccany"), 
   labels = c("Age", "Sex", "Seasonal influenza vaccination"))
   
   table.comp <- descBy(data = df, vars = c("age","sex","fluvaccany"), 
   labels = c("Age", "Sex", "Seasonal influenza vaccination"), 
   compVar = "lab_flu")
}

}
