% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiOption.R
\name{elapsedTime}
\alias{elapsedTime}
\title{elapsedTime}
\usage{
elapsedTime(set = FALSE)
}
\arguments{
\item{set}{TRUE to start}
}
\value{
elapsed time
}
\description{
elapsedTime
}
\examples{
elapsedTime(TRUE)
}
