% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandVar.R
\name{expandVar}
\alias{expandVar}
\title{expandVar}
\usage{
expandVar(data, varname, valueslist)
}
\arguments{
\item{data}{A dataset which contain a multi-variable}

\item{varname}{The name of the multi-variable}

\item{valueslist}{List of expected values with names of the variables to be created}
}
\value{
A df with the new variables
}
\description{
explode a variable which contain a list of values into Y/N variables according to the content of the "multi" variable
the list of expected values should be given with the names of the new variables
}
\examples{

data <- data.frame(Id = 1:4 , 
                    Vaccs = c("pfizer,moderna"," ", "pfizer", "moderna"))
brand <- list("pfizer"="pfizer",
             "moderna"="moderna"
              )
data <-  expandVar(data,Vaccs,brand)


}
