% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{externalFile}
\alias{externalFile}
\title{externalFile}
\usage{
externalFile(extfile)
}
\arguments{
\item{extfile}{Filename}
}
\value{
The full path to access the founded file
}
\description{
This function is used mainly for vignettes. The file is searched
\itemize{
\item locally
\item into extdata from package
\item into inst/extdata locally
}
}
\examples{
externalFile("testbrand.csv")
}
