% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorUsing.R
\name{factorUsing}
\alias{factorUsing}
\title{factorUsing}
\usage{
factorUsing(data, varname, dicoCode, dicoLabel)
}
\arguments{
\item{data}{DataSet which contain \code{varname}}

\item{varname}{The variable to factorise}

\item{dicoCode}{The list of code}

\item{dicoLabel}{The corresponding list of labels}
}
\value{
factored variable
}
\description{
factorUsing
}
