% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorUsingDico.R
\name{factorUsingDico}
\alias{factorUsingDico}
\title{Title}
\usage{
factorUsingDico(data, dictionary = NULL, dicos = NULL)
}
\arguments{
\item{data}{}

\item{dictionary}{}

\item{dicos}{}
}
\value{

}
\description{
Title
}
