% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{fileExt}
\alias{fileExt}
\title{fileExt
Retrieve the file extension}
\usage{
fileExt(text)
}
\arguments{
\item{text}{A filename}
}
\value{
The file extension as character
}
\description{
fileExt
Retrieve the file extension
}
\examples{

myext <- fileExt("mydoc.pdf")

}
