% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{fileFindReplace}
\alias{fileFindReplace}
\title{Find and Replace Text in a File}
\usage{
fileFindReplace(
  filename,
  pattern,
  replacement,
  wholeword = TRUE,
  ignore.case = FALSE
)
}
\arguments{
\item{filename}{The name of the file to modify.}

\item{pattern}{A vector of patterns to search for in the file.}

\item{replacement}{A vector of replacements for the patterns found.}

\item{wholeword}{Logical, if TRUE only whole words will be replaced.}

\item{ignore.case}{Logical, if TRUE the function ignores case.}
}
\value{
No return value; the function modifies the file in-place.
}
\description{
This function reads a file and replaces specified patterns with given replacements.
The function can work with multiple patterns and replacements, and can also perform case-sensitive or case-insensitive searches.
}
\examples{
\dontrun{
fileFindReplace("example.txt", "oldText", "newText")
fileFindReplace("example.txt", c("oldText1", "oldText2"), "newText")
}
}
