% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{fileName}
\alias{fileName}
\title{fileName}
\usage{
fileName(text)
}
\arguments{
\item{text}{A file name}
}
\value{
the filename without path or extension
}
\description{
fileName
}
\examples{
fileName("briths.rds")
}
