% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{fileName}
\alias{fileName}
\title{Extract File Base Name Without Extension}
\usage{
fileName(text)
}
\arguments{
\item{text}{Character. The full path or URL of the file.}
}
\value{
Character. The base name of the file without the extension.
}
\description{
This function takes a file path or URL and returns the base name of the file without the extension.
}
\examples{
\dontrun{
fileName("path/to/file.txt")
fileName("https://example.com/file.csv")
}
}
