% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{fillCells}
\alias{fillCells}
\title{fillCells}
\usage{
fillCells(onesheet, line, col, ..., names = FALSE, wb = NULL)
}
\arguments{
\item{onesheet}{A sheet object from xlsx package}

\item{line}{The line where to paste value}

\item{col}{The col where to paste value}

\item{...}{List of N values to paste in col "col" to col+N
if contain a data.frame, the dataframe is pasted at the position}

\item{names}{if TRUE names of the dataframe are inserted with the table content}

\item{wb}{An optional wb if not already opened}
}
\value{
nothing
}
\description{
fillCells
}
\examples{
wb <- openxlsx::createWorkbook()
sheetname <- "First Sheet"
openxlsx::addWorksheet(wb, sheetName = sheetname)
num <- 3
denum <- 10
fillCells(sheetname,1,1, num,denum , wb=wb )
mat <- data.frame(Id = 1:3 , Vaccs = c("1", "3", "6"))
fillCells(sheetname,1,1, mat ,wb=wb )

}
