% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{fillCells}
\alias{fillCells}
\title{fillCells}
\usage{
fillCells(onesheet, line, col, ..., names = FALSE)
}
\arguments{
\item{onesheet}{A sheet object from xlsx package}

\item{line}{The line where to paste value}

\item{col}{The col where to paste value}

\item{...}{List of N values to paste in col "col" to col+N
if contain a data.frame, the dataframe is pasted at the position}

\item{names}{if TRUE names of the dataframe are inserted with the table content}
}
\value{
nothing
}
\description{
fillCells
}
\examples{
wb <- xlsx::createWorkbook()
sheet <- xlsx::createSheet(wb,"S1")
rows <- xlsx::createRow(sheet, rowIndex = 1:5)
sheetTitle <-xlsx::createCell(rows, colIndex=1:3)
cells <- openSheet("S1",wb)

fillCells(cells,1,1, "Total :",100  )
num <-  40
denum <- 80
fillCells(cells,1,1, num,num/denum  )
mat <- data.frame(Id = 1:3 , Vaccs = c("1", "3", "6"))


}
