% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{fillimage}
\alias{fillimage}
\title{fillimage}
\usage{
fillimage(
  onesheet,
  image,
  line,
  col,
  wide = 7,
  high = 4,
  unit = "in",
  spec.dpi = 300,
  wb = NULL
)
}
\arguments{
\item{onesheet}{A sheet object from xlsx package}

\item{image}{The filename of a previously saved image}

\item{line}{The line where to paste the image}

\item{col}{The col where to paste the image}

\item{wide}{Size of the image}

\item{high}{Size of the image}

\item{unit}{For wide and high, default is "in" (Inch)}

\item{spec.dpi}{300 by default}

\item{wb}{An optional wb if not already opened}
}
\value{
Nothing
}
\description{
fillimage
}
\examples{
/dontrun{ fillimage(onesheet = cells,image = graphname,line=2, col = 2, wb=template) }
}
