% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{formatCells}
\alias{formatCells}
\title{formatCells}
\usage{
formatCells(onesheet, line, col, style = NULL, wb = NULL, ...)
}
\arguments{
\item{onesheet}{A sheet object from xlsx package}

\item{line}{The line where to paste value}

\item{col}{The col where to paste value}

\item{style}{A style created by createXlsxStyle}

\item{wb}{An optional wb if not already opened}

\item{...}{any additional style accedpted by openxlsx::addStyle()}
}
\value{
nothing
}
\description{
formatCells
}
