% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getActionGroup}
\alias{getActionGroup}
\title{getActionGroup}
\usage{
getActionGroup(actiontag)
}
\arguments{
\item{actiontag}{Name of the action group to retrieve}
}
\value{
dataset containing all the variables with actions of type actionname
}
\description{
getActionGroup
}
