% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getAnyDictionaryValue}
\alias{getAnyDictionaryValue}
\title{Title}
\usage{
getAnyDictionaryValue(
  varname,
  search = c("source_name", "generic_name", "dico", "type", "unknowns"),
  value = c("source_name", "generic_name", "dico", "type", "unknowns"),
  dictionary = NULL
)
}
\arguments{
\item{varname}{The varname to search (into the search column)}

\item{search}{The column name which will be searched for varname}

\item{value}{The column name of the value to retrieve}

\item{dictionary}{Optional: specify object to use as dictionary}
}
\value{
One value
}
\description{
Title
}
\examples{
\dontrun{getAnyDictionaryValue("varname",search="source_name",value="dico")}
}
