% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDelayIf.R
\name{getDelayIf}
\alias{getDelayIf}
\title{getDelayIf}
\usage{
getDelayIf(data, FirstDateName, SecondDateName, ...)
}
\arguments{
\item{data}{The dataset containing the values}

\item{FirstDateName}{The first date}

\item{SecondDateName}{The second date}

\item{...}{one or more logical condition to validate. If one of the condition is not TRUE,
NA is returned instead of the calculated delay}
}
\value{
a vector of the same length than the passed dataset
}
\description{
return the delay between two date with condition
if condition are validated then the delay in days from date one to date two is returned
else NA is returned
}
\examples{
\dontrun{
   getDelayIf(data, FirstDateName, SecondDateName, SecondDateName > FirstDateName + 14)
}

}
