% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getDico}
\alias{getDico}
\title{getDico}
\usage{
getDico(diconame, dico = NULL)
}
\arguments{
\item{diconame}{The name of one dico from the dicos structure}

\item{dico}{Optional: specify object to use as dico sheet}
}
\value{
A datset containing one dico (list of code/labels)
}
\description{
getDico
}
