% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getDicoOfVar}
\alias{getDicoOfVar}
\title{getDicoOfVar}
\usage{
getDicoOfVar(varname)
}
\arguments{
\item{varname}{The variable for which we want to retrieve the name of the associated dico}
}
\value{
The name of the dico associated with the variable
}
\description{
getDicoOfVar
}
