% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getDicoOfVar}
\alias{getDicoOfVar}
\title{getDicoOfVar}
\usage{
getDicoOfVar(varname, dictionary = NULL, dico = NULL)
}
\arguments{
\item{varname}{The variable for which we want to retrieve the name of the associated dico}

\item{dictionary}{Optional: specify object to use as dictionary}

\item{dico}{Optional: specify object to use as dico sheet}
}
\value{
The name of the dico associated with the variable
}
\description{
getDicoOfVar
}
