% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getDictionary}
\alias{getDictionary}
\title{getDictionary}
\usage{
getDictionary()
}
\value{
The current dictionary
}
\description{
This function loads the most recent dictionary set into the environment.
}
\examples{
# Create an example data frame to set as a dictionary 
dic <- data.frame(generic_name = c("id", "idcountry", "age", "agegroup3", "agroup10"),
                  source_name = c("id", "idcountry", "age", "agegroup3", "agroup10"),
                  type = c("character", "character", "numeric", "factor", "factor"),
                  dico = c("","","","agegp3","agegp10"))

# Set the current dictionary 
 setDictionary(dictionary = dic)
   
# Load dictionary into the environment 
retrieved <- getDictionary()

}
\seealso{
\code{\link[=setDictionary]{setDictionary()}}
}
