% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getDictionaryValue}
\alias{getDictionaryValue}
\title{getDictionaryValue}
\usage{
getDictionaryValue(varname, valuename = c("type", "dico", "unknowns"))
}
\arguments{
\item{varname}{The varname for which we will retrieve content of one column from dictionary}

\item{valuename}{Name of the coulumn to retrieve from dictionary}
}
\value{
A single value
}
\description{
Retrieve the value of one parameters (column) in the dictionary, searching for the generic_name
usual column values to retrieve are for type, dico, unknowns
Give an error if column name is incorrect
Return NA if searched varname is not found
}
