% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getDictionaryValue}
\alias{getDictionaryValue}
\title{getDictionaryValue}
\usage{
getDictionaryValue(varname, valuename)
}
\arguments{
\item{varname}{The varname for which we will retrieve content of one column from dictionary}

\item{valuename}{Name of the coulumn to retrieve from dictionary}
}
\value{
A single value
}
\description{
getDictionaryValue
}
