% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMath.R
\name{getMax}
\alias{getMax}
\title{getMax}
\usage{
getMax(x)
}
\arguments{
\item{x}{vector}
}
\value{
numeric
}
\description{
get the maximum value from a vector skipping NA values. If all values are NA returns NA.
}
\examples{
getMax(c(1,1))
getMax(c(NA,NA))
}
\seealso{
For more details see the link below to access the vignette:
\href{../doc/epiuf_package.html}{\code{vignette("epiuf_package")}}
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
