% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMath.R
\name{getMax}
\alias{getMax}
\title{getMax}
\usage{
getMax(...)
}
\arguments{
\item{...}{numeric or character vector arguments}
}
\value{
single numeric or character element
}
\description{
Get the maximum value from a vector skipping NA values. If all values are NA, returns NA.
}
\examples{
# Numerics
getMax(c(1, 9))
getMax(1, 9, NA) #c() is not mandatory
getMax(as.Date("2023-01-01"), 
       NA, 
       as.Date("2023-01-31"))

# Characters
getMax(LETTERS) 
## Note: For characters, getMax() will return the last month 
## according to alphabetical order (i.e., "September")
getMax(month.name) 
getMax(NA, month.name)

# If missing values
## using basic max() function from base R
max(c(1:9, NA)) #NA value takes over the other values by default
max(c(1:9, NA), na.rm = TRUE) #Need to add na.rm = TRUE to get the max value despite NAs
max(c(NA, NA), na.rm = TRUE) #If NAs only, returns Inf and a warning
## using getMax, we do not have these issues
getMax(c(NA, 1:9))
getMax(c(NA, NA))

}
\seealso{
\code{\link[base:Extremes]{base::max()}} and \code{\link[=getMin]{getMin()}}, \code{\link[=getMean]{getMean()}},
\code{\link[=getMedian]{getMedian()}} from \code{epiuf} package
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
