% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMath.R
\name{getMin}
\alias{getMin}
\title{getMin}
\usage{
getMin(x)
}
\arguments{
\item{x}{vector}
}
\value{
numeric
}
\description{
get the minimum value from a vector skipping NA values. If all values are NA returns NA.
}
\examples{
getMin(c(1,1))
getMin(c(NA,NA))
}
\seealso{
For more details see the link below to access the vignette:
\href{../doc/epiuf_package.html}{\code{vignette("epiuf_package")}}
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
