% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getNewDictionaryLine}
\alias{getNewDictionaryLine}
\title{Title}
\usage{
getNewDictionaryLine(mode = "dictionary")
}
\arguments{
\item{mode}{Type of line to return. Could be dictionary, dicos, or actions}
}
\value{
An empty record of type \code{mode}
}
\description{
Title
}
