% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSnippets.R
\name{getOS}
\alias{getOS}
\title{getOS}
\usage{
getOS()
}
\value{
character string naming the operator system in use
}
\description{
Returns the user OS in lower case
}
\examples{
getOS()

}
\seealso{
\code{\link[=Sys.info]{Sys.info()}}
}
\author{
Lore Merdrignac
}
