% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{getPath}
\alias{getPath}
\title{getPath}
\usage{
getPath(pathname)
}
\arguments{
\item{pathname}{Path label}
}
\value{
The path saved under pathname label
}
\description{
Retrieve a named Path
}
\examples{
getPath("sources")


}
