% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefSite.R
\name{getRefSiteID}
\alias{getRefSiteID}
\title{Get the ID code corresponding to the country or study site label}
\usage{
getRefSiteID(Label, LocationName)
}
\arguments{
\item{Label}{Character string to search among the "Label" column from epiuf::RefSite reference table}

\item{LocationName}{Character string to search among the "LocationName" column from epiuf::RefSite reference table}
}
\value{
Character string or a data frame if multiple answers
}
\description{
Get the ID code corresponding to the country or study site label
}
\examples{
\dontrun{
getRefSiteID()
}
getRefSiteID(Label = "Austria")
getRefSiteID(LocationName = "Bosnia and Herzegovina")
getRefSiteID(Label = "The Czech Republic", LocationName = "Czechia")
\dontrun{
getRefSiteID(Label = "Czechia")
}

}
\seealso{
For more details see the link below to access the vignette:
\href{../doc/epiuf_package.html}{\code{vignette("epiuf_package")}}
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
