% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefSite.R
\name{getRefSiteLabel}
\alias{getRefSiteLabel}
\title{Get the label corresponding to the ID code of the country or study site label}
\usage{
getRefSiteLabel(ID)
}
\arguments{
\item{ID}{Character string to search among the "ID" column from epiuf::RefSite reference table}
}
\value{
Character string or a data frame if multiple answers
}
\description{
Get the label corresponding to the ID code of the country or study site label
}
\examples{
\dontrun{
getRefSiteLabel()
}
getRefSiteLabel(ID = "SI")
getRefSiteLabel(ID = "UK")
\dontrun{
getRefSiteLabel(ID = "es")
}

}
\seealso{
For more details see the link below to access the vignette:
\href{../doc/epiuf_package.html}{\code{vignette("epiuf_package")}}
}
\author{
STRAP team \email{strap@epiconcept.fr}
}
