% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getVarAction}
\alias{getVarAction}
\title{getVarAction}
\usage{
getVarAction(variablename, actiontag, action = NULL)
}
\arguments{
\item{variablename}{The variable for which we want to retrieve the the associated action}

\item{actiontag}{The name of the action group to retrieve}

\item{action}{Optional: specify object to use as action sheet}
}
\value{
A dataset of var actions records for the variable
}
\description{
getVarAction
}
