% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getVarActionParameters}
\alias{getVarActionParameters}
\title{getVarActionParameters}
\usage{
getVarActionParameters(variablename, actiontag)
}
\arguments{
\item{variablename}{The variable (character) for which we want to retrieve
the associated action parameters (see columns \code{variable} in \code{getDictionaryActions()}).}

\item{actiontag}{The name of the action group (character)
(see column \code{action_group} in \code{getDictionaryActions()}).}
}
\value{
The parameters associated to the variable/action (see column \code{parameters} in \code{getDictionaryActions()})
}
\description{
Retrieve the associated action parameter (i.e., value in \code{parameters} column)
from a given variable (i.e., defined in the column \code{variable} in \code{getDictionaryActions()})
and given action (i.e., value in \code{action_group} column).
}
\seealso{
\code{\link[=getDictionaryActions]{getDictionaryActions()}}, \code{\link[=getNewDictionaryLine]{getNewDictionaryLine()}}, \code{\link[=getVarAction]{getVarAction()}}
}
