% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getVarActionParameters}
\alias{getVarActionParameters}
\title{Title}
\usage{
getVarActionParameters(variablename, actiontag)
}
\arguments{
\item{variablename}{The variable for which we want to retrieve the the associated action parameters}

\item{actiontag}{Name of the action group}
}
\value{

}
\description{
Title
}
