% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{getVarActionParameters}
\alias{getVarActionParameters}
\title{Title}
\usage{
getVarActionParameters(variablename, actiontag)
}
\arguments{
\item{variablename}{}

\item{actiontag}{}
}
\value{

}
\description{
Title
}
