% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{isoWeek}
\alias{isoWeek}
\title{isoWeek}
\usage{
isoWeek(date)
}
\arguments{
\item{date}{A date Value}
}
\value{
The iso Week Number
}
\description{
Returns the ISO week number in numeric format of a date object
}
\examples{
isoWeek(Sys.Date())
isoWeek(as.Date("2020-12-31"))

}
\seealso{
\code{\link[=isoYear]{isoYear()}}, \code{\link[=isoYearWeek]{isoYearWeek()}}  and \code{\link[base:format]{base::format()}}
}
