% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_month_year_tools.R
\name{isoWeek}
\alias{isoWeek}
\title{isoWeek}
\usage{
isoWeek(date)
}
\arguments{
\item{date}{A date Value}
}
\value{
The iso Week Number
}
\description{
isoWeek
}
\examples{
isoWeek(Sys.Date())
}
