% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{isoYear}
\alias{isoYear}
\title{isoYear}
\usage{
isoYear(date)
}
\arguments{
\item{date}{A date value}
}
\value{
The ISO year of the date value which may be different from the
calendar Year
}
\description{
Extract the ISO year of a date object and return it as numeric
}
\examples{
isoYear(Sys.Date())
isoYear(as.Date("2019-12-31"))

}
\seealso{
\code{\link[=calYear]{calYear()}}, \code{\link[=isoYearWeek]{isoYearWeek()}}, \code{\link[=isoWeek]{isoWeek()}}
and \code{\link[base:format]{base::format()}}
}
