% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_month_year_tools.R
\name{isoYear}
\alias{isoYear}
\title{isoYear}
\usage{
isoYear(date)
}
\arguments{
\item{date}{A date value}
}
\value{
The isoYear of the date value which may be different from the
calendar Year
}
\description{
isoYear
}
\examples{
isoYear(Sys.Date())
}
