% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{isoYearWeek}
\alias{isoYearWeek}
\title{isoYearWeek}
\usage{
isoYearWeek(date, weekformat = NULL)
}
\arguments{
\item{date}{A date Value}

\item{weekformat}{A string to represent the week separator. Default is "w".
May be one or more characters (eg : "-W")}
}
\value{
The iso year week of the Date  in YYYYwXX format  (2020w03)
}
\description{
Returns the ISO year and week in character string format (e.g., 2020w03) of a date object
}
\examples{
isoYearWeek(Sys.Date())
isoYearWeek(as.Date("2020-12-31"))

}
\seealso{
\code{\link[=isoYear]{isoYear()}}, \code{\link[=isoWeek]{isoWeek()}}, \code{\link[=lpad]{lpad()}} and \code{\link[base:format]{base::format()}}
}
