% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_month_year_tools.R
\name{isoYearWeek}
\alias{isoYearWeek}
\title{isoYearWeek}
\usage{
isoYearWeek(date, weekformat = NULL)
}
\arguments{
\item{date}{A date Value}

\item{weekformat}{A string to represent the week separator. Default is w.
May be one or more characters (eg : "-W")}
}
\value{
The iso year week of the Date  in YYYYwXX format  (2020w03)
}
\description{
isoYearWeek
}
\examples{
isoYearWeek(Sys.Date())
}
