% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_month_year_tools.R
\name{isoYearWeek}
\alias{isoYearWeek}
\title{isoYearWeek}
\usage{
isoYearWeek(date)
}
\arguments{
\item{date}{A date Value}
}
\value{
The iso year week of the Date  in xx w yyyy format  (43w2020)
}
\description{
isoYearWeek
}
\examples{
isoYearWeek(Sys.Date())
}
