% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_month_year_tools.R
\name{lastDateMonth}
\alias{lastDateMonth}
\title{lastDateMonth}
\usage{
lastDateMonth(month)
}
\arguments{
\item{month}{abbreviated month+year (eg. dec2022)}
}
\value{
The last date in that month in date format
}
\description{
lastDateMonth
}
\examples{
lastDateMonth("dec2022")
}
