% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{left}
\alias{left}
\title{Left}
\usage{
left(text, num_char)
}
\arguments{
\item{text}{Text to extract from}

\item{num_char}{number of char to extract}
}
\value{
\code{num_char} extracted characters from left side
}
\description{
Left
}
\examples{
left("dummy_test",4)

}
