% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFiles.R
\name{listFiles}
\alias{listFiles}
\title{listFiles}
\usage{
listFiles(
  path = ".",
  pattern = "*",
  fullnames = FALSE,
  recursive = FALSE,
  regex = FALSE
)
}
\arguments{
\item{path}{Path to search for files  "." mean current directory}

\item{pattern}{Pattern of files to be listed}

\item{fullnames}{Should the function return full name with path or only file name}

\item{recursive}{If TRUE sub directories are explored (fullnames set to TRUE)}

\item{regex}{If FALSE, the default, the pattern is considered as global pattern
with only standard * and ? jokers}
}
\value{
Character list of found files
}
\description{
listFiles
}
\examples{
listFiles(path = ".", pattern = "*.doc")

}
