% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{listFiles}
\alias{listFiles}
\title{listFiles}
\usage{
listFiles(path = ".", pattern = "*")
}
\arguments{
\item{path}{Path to search for files  "." mean current directory}

\item{pattern}{Pattern of files to be listed}
}
\value{
Character list of found files
}
\description{
listFiles
}
\examples{
listFiles(".", "*.doc")

}
