% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lpad}
\alias{lpad}
\title{lpad
Used to display value with a fixed width format
format value according to width and digit is value is a number}
\usage{
lpad(value, width = 11, digit = 0)
}
\arguments{
\item{value}{A value to format}

\item{width}{The expected width}

\item{digit}{The number of digit}
}
\value{
The formated value
}
\description{
lpad
Used to display value with a fixed width format
format value according to width and digit is value is a number
}
\examples{
lpad("test",10,0)
lpad(2,6,2)
}
