% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDico.R
\name{matchDico}
\alias{matchDico}
\title{matchDico}
\usage{
matchDico(data, mydicos, varname, diconame)
}
\arguments{
\item{data}{The dataset to process}

\item{mydicos}{The dico (epiuf structure) to use}

\item{varname}{The variable to check}

\item{diconame}{Name of the corresponding dico}
}
\value{
A dataset with potential differences
}
\description{
matchDico
}
