% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mid}
\alias{mid}
\title{Extract the midlle of a string}
\usage{
mid(text, start_num, num_char)
}
\arguments{
\item{text}{Text to extract from}

\item{start_num}{start of extraction}

\item{num_char}{number of char to extract}
}
\value{
\code{num_char} extracted characters starting \code{start_num}
}
\description{
Extract \code{num_char} character of a string, starting from the \code{start_num}
character
}
\examples{
mid("dummy_test_2002",7,4)

}
