% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_month_year_tools.R
\name{Month}
\alias{Month}
\title{Month}
\usage{
Month(date)
}
\arguments{
\item{date}{A date Value}
}
\value{
The month of the date in numeric
}
\description{
Month
}
\examples{
Month(Sys.Date())
}
