% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDate.R
\name{Month}
\alias{Month}
\title{Month}
\usage{
Month(date)
}
\arguments{
\item{date}{A date Value}
}
\value{
The month of the date in numeric
}
\description{
Returns the month in numeric format of a date object
}
\examples{
Month(Sys.Date())

}
\seealso{
\code{\link[=abrvMonth]{abrvMonth()}} and \code{\link[base:format]{base::format()}}
}
