% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_drop.R
\name{moyenne}
\alias{moyenne}
\title{Project Name : epiuf package
Script Name  : to_drop
GitHub repo  :
Summary      :
Date created : 2023-08-02
Author       : Lore Merdrignac
Date reviewed:
Reviewed by  :
Description -------------------------------------------------------------}
\usage{
moyenne(x)
}
\arguments{
\item{x}{un vecteur numerique}
}
\value{
la fonction renvoie la moyenne d'un vecteur
}
\description{
Changes Log -------------------------------------------------------------
}
\details{
moyenne d’un vecteur >> Obsolete replaced by getMean

Une fonction pour faire une moyenne en enlevant les valeurs manquantes
}
