% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{openDictionary}
\alias{openDictionary}
\title{Title}
\usage{
openDictionary(filename)
}
\arguments{
\item{filename}{The file (xls) containing a dictionary. The dictionary will be loaded}
}
\value{

}
\description{
Title
}
