% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{openSheet}
\alias{openSheet}
\title{openSheet}
\usage{
openSheet(sheetname, wb = NULL)
}
\arguments{
\item{sheetname}{The name of hte sheet to load}

\item{wb}{A optional workbook created or opened by xlsx
If not specified, the last workbook loaded with loadXlsx will be used}
}
\value{
A cells object containing an Excel sheet
}
\description{
openSheet
}
\examples{
 wb <- xlsx::createWorkbook()
 sheet <- xlsx::createSheet(wb,"S1")
 rows <- xlsx::createRow(sheet, rowIndex = 1:5)
 sheetTitle <-xlsx::createCell(rows, colIndex=1)

 cells <- openSheet("S1",wb)

}
