% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{openSheet}
\alias{openSheet}
\title{openSheet}
\usage{
openSheet(sheetname, wb = NULL)
}
\arguments{
\item{sheetname}{The name of hte sheet to load}

\item{wb}{A optional workbook created or opened by xlsx
If not specified, the last workbook loaded with loadXlsx will be used}
}
\value{
A cells object containing an Excel sheet
}
\description{
openSheet
}
\examples{
 wb <- openxlsx::createWorkbook()
 openxlsx::addWorksheet(wb, sheetName = "First Sheet")
 openSheet("First Sheet",wb)

}
