% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{openXlsx}
\alias{openXlsx}
\title{loadXlsx}
\usage{
openXlsx(filename = "")
}
\arguments{
\item{filename}{Name of the Excel file}
}
\value{
The wb (which is also saved as internal variable)
}
\description{
loadXlsx
}
\examples{
cat("to be done")

}
