% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{pathToFile}
\alias{pathToFile}
\title{pathToFile}
\usage{
pathToFile(pathname, filename)
}
\arguments{
\item{pathname}{Label for saved path}

\item{filename}{Name of the file to retreive, may contain subdir}
}
\value{
the full name of the file
}
\description{
construct an absolute path to a file using file name and saved/labeled path
}
\examples{
pathToFile("data","cleanedrecords.csv")


}
