% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pos}
\alias{pos}
\title{Retrieve the position of a char in a string}
\usage{
pos(pattern, stosearch)
}
\arguments{
\item{pattern}{A pattern to search in stosearch}

\item{stosearch}{A character string}
}
\value{
Position of pattern in stosearch
}
\description{
pos is used to find the position of a specific char or pattern into a string
If pattern is a single character, the first position of this character in the searched
string is returned
If pattern is a regular expression, the result is 0 if the expression is not verified
or the position of the first pattern in the string \code{stosearch}
}
\examples{
pos("/","test/string")

}
