% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printDataCheck.R
\name{printDataCheck}
\alias{printDataCheck}
\title{Title}
\usage{
printDataCheck(
  data,
  data_old = NULL,
  parameters,
  header = "",
  threshold = NULL,
  varname = "id"
)
}
\arguments{
\item{data}{The dataset}

\item{data_old}{A previously check dataset}

\item{parameters}{...}

\item{header}{The header to display before list of ID}

\item{threshold}{The maximum number of ID to be displayed}

\item{varname}{The ID varname}
}
\value{
data.frame
}
\description{
Title
}
\examples{
\dontrun{   printDataCheck(data=df, data_old=NULL,  parameters = parameters, 
                  header="Essential Data Checks", threshold=NULL , varname="idCheck") }
                  
}
