% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printIfComp.R
\name{printIfComp}
\alias{printIfComp}
\title{printIfComp}
\usage{
printIfComp(data, data_old, cond, text = "", threshold = NULL, varname = "id")
}
\arguments{
\item{data}{The dataset to check}

\item{data_old}{The other dataset which may contain same records}

\item{cond}{A condition to meet}

\item{text}{A message to display if condition is not meet}

\item{threshold}{Something ..}

\item{varname}{The ID variaable to display}
}
\value{
The text to print
}
\description{
printIfComp
}
