% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printIfComp.R
\name{printIfComp}
\alias{printIfComp}
\title{Title}
\usage{
printIfComp(data, data_old, cond, text = "", threshold = NULL, varname = "id")
}
\arguments{
\item{data}{}

\item{data_old}{}

\item{cond}{}

\item{text}{}

\item{threshold}{}

\item{varname}{}
}
\value{

}
\description{
Title
}
