% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printVar}
\alias{printVar}
\title{printVar}
\usage{
printVar(dataset, pattern, regex = FALSE)
}
\arguments{
\item{dataset}{A data.frame}

\item{pattern}{A pattern for varname in the dataframe see \code{regex}}

\item{regex}{Should the pattern be used as regex expression or use classical "joker" ? and *}
}
\value{
A data frame of 10 rows with selected columns
}
\description{
printVar
}
\examples{
 data <- data.frame(Id = 1:4 ,  
                   vaccage = c(34,45, 50,22 ),
                   symp = c("Y","Y","N","N"),
                   vaccboost=c("N","Y","N","Y"))
printVar(data,"symp")
printVar(data,"vac*")                    
}
