% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Read Data from Various File Types}
\usage{
readData(
  filename = "",
  factorise = FALSE,
  lowercase = FALSE,
  label = NULL,
  sheet = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{filename}{String specifying the path of the file to read. If empty, opens a file dialog.}

\item{factorise}{Logical, if TRUE factorizes the data.}

\item{lowercase}{Logical, if TRUE converts column names to lowercase.}

\item{label}{Optional label to add to the data.}

\item{sheet}{Sheet name or index for Excel files (If reading an Excel file)}

\item{verbose}{Logical, if TRUE prints details about the loaded file.}

\item{...}{Additional arguments passed to the underlying read functions
like encoding="latin1".}
}
\value{
Returns a data frame of the loaded file, or an error if the file format
is not supported or the file does not exist.
}
\description{
This function reads data from various file types such as CSV, DTA (Stata), REC (EpiInfo), RDA/RDATA, RDS, and Excel files.
It provides options to factorize, convert column names to lowercase, and add labels to the data.
If no filename is provided, a file dialog will be opened for the user to select a file.
}
\examples{
readData("flucases.csv")


}
