% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReverseFactorUsingDico.R
\name{reverseFactorUsingDico}
\alias{reverseFactorUsingDico}
\title{Title}
\usage{
reverseFactorUsingDico(data, dictionary = NULL, dicos = NULL)
}
\arguments{
\item{data}{}

\item{dictionary}{}

\item{dicos}{}
}
\value{

}
\description{
Title
}
