% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{saveDictionary}
\alias{saveDictionary}
\title{Title}
\usage{
saveDictionary(filename = NULL, dictionary = NULL)
}
\arguments{
\item{filename}{}

\item{dictionary}{}
}
\value{

}
\description{
Title
}
