% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{saveDictionary}
\alias{saveDictionary}
\title{saveDictionary}
\usage{
saveDictionary(filename = NULL, dictionary = NULL)
}
\arguments{
\item{filename}{The filename where to save the dictionary.
This file will be erased and replaced by a new  one containing the \code{dictionary}}

\item{dictionary}{Optionaly a dictionary. By default the current dictionary will be saved}
}
\value{
Nothing
}
\description{
saveDictionary
}
