% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{saveXlsx}
\alias{saveXlsx}
\title{saveXlsx
This function save the current workbook as filename}
\usage{
saveXlsx(filename = "", wb = NULL)
}
\arguments{
\item{filename}{An optional filename (if empty the current filename will be used )}

\item{wb}{A optional workbook created or opened by xlsx
If not specified, the last workbook loaded with loadXlsx will be used}
}
\value{
nothing
}
\description{
saveXlsx
This function save the current workbook as filename
}
\examples{

# Creating an empty workbook
wb <- openxlsx::createWorkbook()
sheetname <- "First Sheet"
openxlsx::addWorksheet(wb, sheetName = sheetname)

# Writing a data frame in the workbook 
fillCells(wb = wb, 
          onesheet = sheetname, 
          line = 1, 
          col = 1, 
          iris)
          
# Saving the Excel file
\dontrun{
saveXlsx(filename = "test.xlsx", wb = wb)
}

}
