% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsxTools.R
\name{saveXlsx}
\alias{saveXlsx}
\title{saveXlsx
This function save the current workbook as filename}
\usage{
saveXlsx(filename = "", wb = NULL)
}
\arguments{
\item{filename}{An optional filename (if empty the current filename will be used )}

\item{wb}{A optional workbook created or opened by xlsx
If not specified, the last workbook loaded with loadXlsx will be used}
}
\value{
nothing
}
\description{
saveXlsx
This function save the current workbook as filename
}
\examples{

cat("to be done")

}
