% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{setDicos}
\alias{setDicos}
\title{Title}
\usage{
setDicos(dic)
}
\arguments{
\item{dic}{A dataset of dicos (with epiuf structure as returned by getNewDictionaryLine)}
}
\value{

}
\description{
Title
}
