% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{setDictionary}
\alias{setDictionary}
\title{setDictionary}
\usage{
setDictionary(dictionary)
}
\arguments{
\item{dictionary}{A dictionary epiuf structure (data.frame)}
}
\value{
Nothing
}
\description{
Set a dictionary using a data.frame.
}
\examples{

# Create an example data frame to set as a dictionary 
dic <- data.frame(generic_name=c("gen1","gen2","gen3","gen4","gen5",NA,NA),
  source_name=c("source1","source2","source3","",NA,"source6",NA),
  type=c("numeric","numeric","numeric","character"," ","character",NA),
  unknowns=c("","8","UNK","","","",""))

# Set the dictionary 
dictionary <- setDictionary(dictionary = dic)


}
