% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{setDictionary}
\alias{setDictionary}
\title{setDictionary}
\usage{
setDictionary(dictionary)
}
\arguments{
\item{dictionary}{A dictionary epiuf structure}
}
\value{
Nothing
}
\description{
setDictionary
}
