% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applydictionary.R
\name{setDictionaryActions}
\alias{setDictionaryActions}
\title{setDictionaryActions}
\usage{
setDictionaryActions(actions)
}
\arguments{
\item{actions}{A dataset of actions}
}
\value{
Nothing
}
\description{
setDictionaryActions
}
