% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{setPath}
\alias{setPath}
\title{setPath}
\usage{
setPath(pathname, path, makedir = c("Ask", "Force", "Never"))
}
\arguments{
\item{pathname}{Pathname label specified in "".}

\item{path}{Path to the desired folder or file that you want linked with the pathname. See function 'pathToFile()'.}

\item{makedir}{If "Force" path will be created if it doesn't exist, if "Never" path will not be created and
a warning will pop if path is missing. If set to "Ask" a prompt will ask for confirmation before
creating the missing directory.}
}
\value{
Previous defined path
}
\description{
Set a named path to avoid absolute path in R scripts. Creates a short cut to folders and files.
}
\examples{

\dontrun{
# Set a path named 'sources' 
setPath(pathname = "SOURCES", 
    path = "c:/dev/Resources", 
    makedir = "Never")
# Check path is properly set 
getPath("SOURCES")

# Setting a name/keyword associated to a specific path
setPath(pathname = "DATA", 
        path = "./data", # Note './' for relative paths
        makedir = "Never")
# Checking the path is properly set
getPath("DATA")
}


}
\seealso{
\code{\link[=pathToFile]{pathToFile()}} and \code{\link[=getPath]{getPath()}}
}
