% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{setPath}
\alias{setPath}
\title{setPath}
\usage{
setPath(pathname, path, makedir = c("Ask", "Force", "Never"))
}
\arguments{
\item{pathname}{pathname Label}

\item{path}{absolute path to associate with the pathname}

\item{makedir}{if "Force" path will be created if it doesn't exist, if "Never" path will not be created and
a warning will pop if path is missing. If set to "Ask" a prompt will ask for confirmation before
creating the missing directory}
}
\value{
Previous defined path
}
\description{
Set a named path to avoid absolute path in R scripts
}
\examples{

setPath("sources","c:/dev/Rsources", makedir = "Never")

}
