% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{setPath}
\alias{setPath}
\title{setPath}
\usage{
setPath(pathname, path)
}
\arguments{
\item{pathname}{pathname Label}

\item{path}{absolute path to associate with the pathname}
}
\value{
Previous defined path
}
\description{
Set a named path to avoid absolute path in R scripts
}
\examples{

setPath("sources","c:/dev/Rsources")

}
