% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setType.R
\name{setType}
\alias{setType}
\title{setType
Transforms a column in a dataset to the defined class (character, date, factor,
numeric)}
\usage{
setType(data, varname, type = c("character", "date", "factor", "numeric"))
}
\arguments{
\item{data}{The dataset}

\item{varname}{The var}

\item{type}{The desired type}
}
\value{
variable
}
\description{
setType
Transforms a column in a dataset to the defined class (character, date, factor,
numeric)
}
\examples{
\dontrun{
   setType(data,var, "character")
}

}
