% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTypeAll.R
\name{setTypeAll}
\alias{setTypeAll}
\title{setTypeAll
Uses dictionary to define the class for each variable.}
\usage{
setTypeAll(data, dictionary = NULL)
}
\arguments{
\item{data}{The dataset}

\item{dictionary}{The epiuf dictionary}
}
\value{
data frame
}
\description{
setTypeAll
Uses dictionary to define the class for each variable.
}
\examples{
 \dontrun{
   setTypeAll(data)
   }

}
