% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortby.R
\name{sortBy}
\alias{sortBy}
\title{sortBy
Sort a dataframe according to a field optionnaly decreasing}
\usage{
sortBy(data, field, decreasing = FALSE, na.last = TRUE)
}
\arguments{
\item{data}{The dataframe to besorted}

\item{field}{The field on which to sort (numeric or character).}

\item{decreasing}{Default is FALSE}

\item{na.last}{Should NA be at the end ?}
}
\value{
The sorted dataframe
}
\description{
sortBy
Sort a dataframe according to a field optionnaly decreasing
}
\examples{
df <- data.frame(one=c(2,1,2,1),two=c(2,2,1,1))
df <- sortBy(df,"one")
df <- sortBy(df,1) 
}
