% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{sourceFile}
\alias{sourceFile}
\title{sourceFile}
\usage{
sourceFile(pathname, filename)
}
\arguments{
\item{pathname}{Label attributed to the path where to find the file to source,
specified in character string (e.g., \code{"SCRIPTS"}).}

\item{filename}{The name of the file to source in character string
(if it contains a path, it will be added to the specified path (\code{pathname}).}
}
\value{
No return value
}
\description{
Source a file from a previously set path and run it.
It uses the label of the path (\code{pathname}) and the name of the file
(\code{filename}) of which you want to run.
This function is a wrapper for 'source()' base R function.
}
\examples{
\dontrun{
# Set an example path named 'SCRIPTS' 
setPath(pathname = "SCRIPTS", 
       path = "./scripts", 
       makedir = "Never")

# Source an R script from the specified path 
sourceFile("SCRIPTS", "import_data.R")
}

}
\seealso{
\code{\link[=setPath]{setPath()}}
}
