% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{sourceFile}
\alias{sourceFile}
\title{sourceFile}
\usage{
sourceFile(pathname, filename)
}
\arguments{
\item{pathname}{pathname as label}

\item{filename}{name of the file to source (if it contain a path, this one will be added to the scripts path)}
}
\value{
nothing
}
\description{
sourcefile function is a wrapper for source which will later allow some test and feebacks
sourceFile source a file from the "scripts" path as set by setPath("scripts","absolute path")
}
\examples{
file <- tempfile(fileext = ".R")
cat("# comment", file = file, sep = "\n")
setPath("scripts","")
sourceFile("scripts",file)
unlink(file) # tidy up

}
