% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{tab}
\usage{
tab(..., includeNA = "ifany", data)
}
\arguments{
\item{...}{Variable or vector}

\item{includeNA}{Should NA be included in results ?}

\item{data}{The dataset which contain the variables}
}
\value{
cross table
}
\description{
tab
}
