% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validDate.R
\name{validDate}
\alias{validDate}
\title{transforms any character into a valid R Date format}
\usage{
validDate(datevar, format = NULL)
}
\arguments{
\item{datevar}{an (character) array to convert it to a Date array}

\item{format}{an optional format string to use in place of automatic system, useful for some complex format}
}
\value{
an (Date) array already converted into the Universal Date Format (UDF) in R (YYYY-mm-dd)
}
\description{
transforms any character into a valid R Date format
}
\examples{
datevar <- c("2022/01/15","2022/02/10","1950/01/31","2022/15/15",NA)
datevar <- c("20220115","20220210","19500131","111",NA)

}
