% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyspelling.R
\name{verifySpelling}
\alias{verifySpelling}
\title{Title}
\usage{
verifySpelling(varname, CorrectList, ErrPerc = 0.25)
}
\arguments{
\item{varname}{a string which will be compraed to anotehr string or a list of strings}

\item{CorrectList}{a string or a list of string from which varname will be searched}

\item{ErrPerc}{An acceptable errPerc when comparing string , default to 0.10\%}
}
\value{
The string guessed from varname using CorrectList
}
\description{
Title
}
\examples{

verifySpelling("Janury",c("January", "Janvier" ))

}
