% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewIf.R
\name{viewIf}
\alias{viewIf}
\title{List head of data.frame allowing to select column to view}
\usage{
viewIf(data, ..., cond = NULL, nline = 10)
}
\arguments{
\item{data}{The dataset to explore}

\item{...}{variable name(s) to include in view}

\item{cond}{A logical condition to select records which will be evaluated into data, given as character string}

\item{nline}{Number of rows to show (default to n) If negative then number of row from the end}
}
\value{
invisibly, the resulting selection
}
\description{
List head of data.frame allowing to select column to view
}
\examples{
data <- data.frame(id = 1:10,
                   case = c(rep(1,3), rep(0,7)),
                   vacc = sample(c(0,1), replace = TRUE, size = 10),
                   ill = sample(c(0,1), replace = TRUE, size = 10))
data[8,2]<-NA                    
viewIf(data ,case,vacc)
viewIf(data, ,case,vacc, nline=3)
viewIf(data,case,ill, nline= -3)
viewIf(data,case,ill,cond="vacc==1" )
 
}
